@@ -50,7 +50,7 @@ multi_line_output = 3
 line_length = 88
 known_first_party = ["src", "tests"]
 known_third_party = [
-    "flask", "qdrant_client", "langchain", "openai", 
+    "flask", "qdrant_client", "langchain", "openai",
     "azure", "google", "msgraph", "clamd", "spamassassin"
 ]
 sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
@@ -104,7 +104,7 @@ python_classes = ["Test*"]
 python_functions = ["test_*"]
 markers = [
     "unit: Unit tests",
-    "integration: Integration tests", 
+    "integration: Integration tests",
     "slow: Slow tests",
     "email: Email interface tests",
     "ai: AI processing tests",
@@ -128,7 +128,7 @@ max-line-length = 88
 [tool.pylint.messages_control]
 disable = [
     "too-few-public-methods",
-    "too-many-arguments", 
+    "too-many-arguments",
     "too-many-instance-attributes",
     "too-many-locals",
     "line-too-long",  # Handled by black
@@ -167,23 +167,28 @@ exclude_dirs = ["tests", ".emailagent", "build", "dist"]
 skips = ["B101", "B601"]  # Skip assert_used and shell_injection for development
 
 [tool.ruff]
-target-version = "py38"
-line-length = 88
+target-version = "py311"
+line-length = 120
+
+[tool.ruff.lint]
 select = [
-    "E",  # pycodestyle errors
-    "W",  # pycodestyle warnings
-    "F",  # pyflakes
-    "I",  # isort
-    "B",  # flake8-bugbear
-    "C4", # flake8-comprehensions
-    "UP", # pyupgrade
+    "E",   # pycodestyle errors
+    "W",   # pycodestyle warnings
+    "F",   # pyflakes
+    "I",   # isort
+    "B",   # flake8-bugbear
+    "C4",  # flake8-comprehensions
+    "UP",  # pyupgrade
+    "SIM", # flake8-simplify
+    "T201", # Disallow bare print statements (flake8-print)
 ]
 ignore = [
-    "E501",  # line too long, handled by black
+    "E501",  # line too long (handled by formatter)
     "B008",  # do not perform function calls in argument defaults
     "C901",  # too complex
 ]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"]  # Allow unused imports in __init__.py
-"tests/**/*" = ["E501"]   # Allow longer lines in tests 
+"tests/**/*" = ["E501", "T201"]   # Allow print in tests
+"scripts/**/*" = ["T201"]           # Allow print in helper scripts
