@@ -57,23 +57,24 @@ Copyright 2025 by Inveniam Capital Partners, LLC and Rick Bunker
 
 import asyncio
 import hashlib
-import uuid
+import os
 import re
-import subprocess
-import tempfile
 import shutil
-from datetime import datetime, timezone
-from pathlib import Path
-from typing import List, Dict, Any, Optional, Tuple, Union, Set
-from dataclasses import dataclass, field
-from enum import Enum
+import subprocess
 
 # Logging system
 import sys
-import os
+import tempfile
+import uuid
+from dataclasses import dataclass
+from datetime import UTC, datetime
+from enum import Enum
+from pathlib import Path
+from typing import Any
+
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 
-from utils.logging_system import get_logger, log_function, log_debug, log_info
+from utils.logging_system import get_logger, log_function
 
 # Initialize logger
 logger = get_logger(__name__)
@@ -96,8 +97,14 @@ except ImportError:
 try:
     from qdrant_client import QdrantClient
     from qdrant_client.models import (
-        CollectionConfig, VectorParams, Distance, 
-        PointStruct, Filter, FieldCondition, MatchValue, PointIdsList
+        CollectionConfig,
+        Distance,
+        FieldCondition,
+        Filter,
+        MatchValue,
+        PointIdsList,
+        PointStruct,
+        VectorParams,
     )
     logger.info("Qdrant client library loaded successfully")
 except ImportError:
@@ -146,27 +153,27 @@ class DocumentCategory(Enum):
     APPRAISAL = "appraisal"
     LEASE_DOCUMENTS = "lease_documents"
     PROPERTY_MANAGEMENT = "property_management"
-    
+
     # Private Credit
     LOAN_DOCUMENTS = "loan_documents"
     BORROWER_FINANCIALS = "borrower_financials"
     COVENANT_COMPLIANCE = "covenant_compliance"
     CREDIT_MEMO = "credit_memo"
     LOAN_MONITORING = "loan_monitoring"
-    
+
     # Private Equity
     PORTFOLIO_REPORTS = "portfolio_reports"
     INVESTOR_UPDATES = "investor_updates"
     BOARD_MATERIALS = "board_materials"
     DEAL_DOCUMENTS = "deal_documents"
     VALUATION_REPORTS = "valuation_reports"
-    
+
     # Infrastructure
     ENGINEERING_REPORTS = "engineering_reports"
     CONSTRUCTION_UPDATES = "construction_updates"
     REGULATORY_DOCUMENTS = "regulatory_documents"
     OPERATIONS_REPORTS = "operations_reports"
-    
+
     # General
     LEGAL_DOCUMENTS = "legal_documents"
     TAX_DOCUMENTS = "tax_documents"
@@ -178,32 +185,32 @@ class ConfidenceLevel(Enum):
     """Confidence levels for routing decisions"""
     HIGH = "high"          # ≥ 90% - Auto-process
     MEDIUM = "medium"      # ≥ 70% - Process with confirmation
-    LOW = "low"           # ≥ 50% - Save uncategorized  
+    LOW = "low"           # ≥ 50% - Save uncategorized
     VERY_LOW = "very_low" # < 50% - Human review required
 
 @dataclass
 class ProcessingResult:
     """Result of document processing"""
     status: ProcessingStatus
-    file_hash: Optional[str] = None
-    file_path: Optional[Path] = None
+    file_hash: str | None = None
+    file_path: Path | None = None
     confidence: float = 0.0
-    error_message: Optional[str] = None
-    quarantine_reason: Optional[str] = None
-    duplicate_of: Optional[str] = None
-    metadata: Dict[str, Any] = None
-    
+    error_message: str | None = None
+    quarantine_reason: str | None = None
+    duplicate_of: str | None = None
+    metadata: dict[str, Any] = None
+
     # Phase 3: Document Classification
-    document_category: Optional[DocumentCategory] = None
-    confidence_level: Optional[ConfidenceLevel] = None
-    matched_asset_id: Optional[str] = None
+    document_category: DocumentCategory | None = None
+    confidence_level: ConfidenceLevel | None = None
+    matched_asset_id: str | None = None
     asset_confidence: float = 0.0
-    classification_metadata: Dict[str, Any] = None
+    classification_metadata: dict[str, Any] = None
 
 @dataclass
 class AssetDocumentConfig:
     """Configuration for asset document types"""
-    allowed_extensions: List[str]
+    allowed_extensions: list[str]
     max_file_size_mb: int
     quarantine_days: int
     version_retention_count: int
@@ -216,10 +223,10 @@ class Asset:
     asset_name: str  # Full descriptive name
     asset_type: AssetType
     folder_path: str  # File system path
-    identifiers: List[str]  # Alternative names/identifiers
+    identifiers: list[str]  # Alternative names/identifiers
     created_date: datetime
     last_updated: datetime
-    metadata: Dict[str, Any] = None
+    metadata: dict[str, Any] = None
 
 @dataclass
 class AssetSenderMapping:
@@ -228,11 +235,11 @@ class AssetSenderMapping:
     asset_id: str  # Asset deal_id
     sender_email: str
     confidence: float  # 0.0 to 1.0
-    document_types: List[str]  # Expected document categories
+    document_types: list[str]  # Expected document categories
     created_date: datetime
     last_activity: datetime
     email_count: int = 0
-    metadata: Dict[str, Any] = None
+    metadata: dict[str, Any] = None
 
 @dataclass
 class UnknownSender:
@@ -241,10 +248,10 @@ class UnknownSender:
     first_seen: datetime
     last_seen: datetime
     email_count: int
-    pending_documents: List[str]  # File hashes
+    pending_documents: list[str]  # File hashes
     timeout_hours: int = 48
     escalated: bool = False
-    metadata: Dict[str, Any] = None
+    metadata: dict[str, Any] = None
 
 class AssetDocumentAgent:
     """
@@ -254,7 +261,7 @@ class AssetDocumentAgent:
     Phase 1: File validation, hashing, and antivirus scanning
     Phase 2: Asset management, sender mapping, and Qdrant integration
     """
-    
+
     # Asset type configurations
     ASSET_CONFIGS = {
         AssetType.COMMERCIAL_REAL_ESTATE: AssetDocumentConfig(
@@ -282,7 +289,7 @@ class AssetDocumentAgent:
             version_retention_count=10
         )
     }
-    
+
     # Qdrant collection names
     COLLECTIONS = {
         'assets': 'asset_management_assets',
@@ -290,11 +297,11 @@ class AssetDocumentAgent:
         'processed_documents': 'asset_management_processed_documents',
         'unknown_senders': 'asset_management_unknown_senders'
     }
-    
-    def __init__(self, 
-                 qdrant_client: Optional[QdrantClient] = None,
+
+    def __init__(self,
+                 qdrant_client: QdrantClient | None = None,
                  base_assets_path: str = "./assets",
-                 clamav_socket: Optional[str] = None) -> None:
+                 clamav_socket: str | None = None) -> None:
         """
         Initialize the Asset Document Agent.
         
@@ -312,11 +319,11 @@ class AssetDocumentAgent:
         """
         self.logger = get_logger(f"{__name__}.AssetDocumentAgent")
         self.logger.info("Initializing Asset Document Agent")
-        
+
         # Core components
         self.qdrant = qdrant_client
         self.base_assets_path = Path(base_assets_path)
-        
+
         # Create base directory
         try:
             self.base_assets_path.mkdir(parents=True, exist_ok=True)
@@ -324,27 +331,27 @@ class AssetDocumentAgent:
         except OSError as e:
             self.logger.error(f"Failed to create assets directory: {e}")
             raise
-        
+
         # Initialize ClamAV
-        self.clamscan_path: Optional[str] = None
+        self.clamscan_path: str | None = None
         self._init_clamav(clamav_socket)
-        
+
         # Processing statistics
-        self.stats: Dict[str, int] = {
+        self.stats: dict[str, int] = {
             'processed': 0,
             'quarantined': 0,
             'duplicates': 0,
             'errors': 0
         }
-        
+
         # Performance metrics
-        self.processing_times: List[float] = []
-        self.classification_cache: Dict[str, Tuple[DocumentCategory, float]] = {}
-        
+        self.processing_times: list[float] = []
+        self.classification_cache: dict[str, tuple[DocumentCategory, float]] = {}
+
         self.logger.info("Asset Document Agent initialized successfully")
-        
+
     @log_function()
-    def _init_clamav(self, socket_path: Optional[str] = None) -> None:
+    def _init_clamav(self, socket_path: str | None = None) -> None:
         """
         Initialize ClamAV antivirus scanning capabilities.
         
@@ -359,7 +366,7 @@ class AssetDocumentAgent:
             compatibility across different deployment environments.
         """
         self.logger.info("Initializing ClamAV antivirus integration")
-        
+
         # Common ClamAV installation paths
         clamscan_paths = [
             '/opt/homebrew/bin/clamscan',  # macOS Homebrew
@@ -367,13 +374,13 @@ class AssetDocumentAgent:
             '/usr/local/bin/clamscan',     # Other Unix systems
             'C:\\Program Files\\ClamAV\\clamscan.exe',  # Windows
         ]
-        
+
         self.clamscan_path = None
-        
+
         # Method 1: Try to find clamscan using which/where command
         try:
             which_cmd = 'where' if os.name == 'nt' else 'which'
-            result = subprocess.run([which_cmd, 'clamscan'], 
+            result = subprocess.run([which_cmd, 'clamscan'],
                                   capture_output=True, text=True, timeout=5)
             if result.returncode == 0 and result.stdout.strip():
                 self.clamscan_path = result.stdout.strip()
@@ -381,14 +388,14 @@ class AssetDocumentAgent:
                 return
         except (subprocess.TimeoutExpired, FileNotFoundError) as e:
             self.logger.debug(f"which/where command failed: {e}")
-        
+
         # Method 2: Check common installation paths
         for path in clamscan_paths:
             if Path(path).exists():
                 self.clamscan_path = path
                 self.logger.info(f"Found clamscan at standard path: {self.clamscan_path}")
                 return
-        
+
         # Method 3: Use shutil.which as final fallback
         self.clamscan_path = shutil.which('clamscan')
         if self.clamscan_path:
@@ -396,7 +403,7 @@ class AssetDocumentAgent:
         else:
             self.logger.warning("ClamAV clamscan not found - antivirus scanning disabled")
             self.logger.info("Install ClamAV: brew install clamav (macOS) or apt-get install clamav (Ubuntu)")
-            
+
         # Validate clamscan functionality if found
         if self.clamscan_path:
             self._validate_clamscan()
@@ -414,14 +421,14 @@ class AssetDocumentAgent:
         """
         if not self.clamscan_path:
             return False
-            
+
         try:
             # Test clamscan with version check
             result = subprocess.run(
                 [self.clamscan_path, '--version'],
                 capture_output=True, text=True, timeout=10
             )
-            
+
             if result.returncode == 0:
                 version_info = result.stdout.strip()
                 self.logger.info(f"ClamAV validation successful: {version_info}")
@@ -429,7 +436,7 @@ class AssetDocumentAgent:
             else:
                 self.logger.error(f"ClamAV validation failed: {result.stderr}")
                 return False
-                
+
         except Exception as e:
             self.logger.error(f"ClamAV validation error: {e}")
             return False
@@ -453,13 +460,13 @@ class AssetDocumentAgent:
         """
         if not file_content:
             raise ValueError("File content cannot be empty or None")
-            
+
         file_hash = hashlib.sha256(file_content).hexdigest()
         self.logger.debug(f"Calculated SHA256 hash: {file_hash[:16]}...")
         return file_hash
-    
+
     @log_function()
-    def validate_file_type(self, filename: str, asset_type: Optional[AssetType] = None) -> bool:
+    def validate_file_type(self, filename: str, asset_type: AssetType | None = None) -> bool:
         """
         Validate file type against allowed extensions.
         
@@ -478,10 +485,10 @@ class AssetDocumentAgent:
         """
         if not filename:
             raise ValueError("Filename cannot be empty or None")
-            
+
         file_extension = Path(filename).suffix.lower()
         self.logger.debug(f"Validating file type: {filename} (extension: {file_extension})")
-        
+
         if asset_type:
             config = self.ASSET_CONFIGS[asset_type]
             is_valid = file_extension in config.allowed_extensions
@@ -489,14 +496,14 @@ class AssetDocumentAgent:
             return is_valid
         else:
             # Check against all asset types
-            all_allowed: Set[str] = set()
+            all_allowed: set[str] = set()
             for config in self.ASSET_CONFIGS.values():
                 all_allowed.update(config.allowed_extensions)
             is_valid = file_extension in all_allowed
             self.logger.debug(f"File type validation against all asset types: {is_valid}")
             return is_valid
-    
-    def validate_file_size(self, file_size: int, asset_type: Optional[AssetType] = None) -> bool:
+
+    def validate_file_size(self, file_size: int, asset_type: AssetType | None = None) -> bool:
         """
         Validate file size against limits.
         
@@ -510,10 +517,10 @@ class AssetDocumentAgent:
         else:
             # Use maximum allowed size across all asset types
             max_size = max(config.max_file_size_mb for config in self.ASSET_CONFIGS.values()) * 1024 * 1024
-        
+
         return file_size <= max_size
-    
-    async def scan_file_antivirus(self, file_content: bytes, filename: str) -> Tuple[bool, Optional[str]]:
+
+    async def scan_file_antivirus(self, file_content: bytes, filename: str) -> tuple[bool, str | None]:
         """
         Scan file content with ClamAV using command-line clamscan.
         
@@ -521,17 +528,15 @@ class AssetDocumentAgent:
             (is_clean, threat_name)
         """
         if not self.clamscan_path:
-            print(f"⚠️ ClamAV not available - skipping scan for {filename}")
+            self.logger.warning("ClamAV not available – skipping scan for %s", filename)
             return True, None
-        
-        import tempfile
-        import subprocess
-        
+
+
         # Save file content to a temporary file
         with tempfile.NamedTemporaryFile(delete=False) as temp_file:
             temp_file.write(file_content)
             temp_path = temp_file.name
-        
+
         try:
             # Run clamscan asynchronously on the file
             process = await asyncio.create_subprocess_exec(
@@ -539,15 +544,15 @@ class AssetDocumentAgent:
                 stdout=asyncio.subprocess.PIPE,
                 stderr=asyncio.subprocess.PIPE
             )
-            
+
             stdout, stderr = await asyncio.wait_for(
                 process.communicate(), timeout=30  # 30 second timeout
             )
-            
+
             # Decode bytes to string
             stdout = stdout.decode('utf-8') if stdout else ""
             stderr = stderr.decode('utf-8') if stderr else ""
-            
+
             # Check if virus was found
             if 'Infected files: 1' in stdout or process.returncode == 1:
                 # Extract virus name from output
@@ -555,27 +560,27 @@ class AssetDocumentAgent:
                     if temp_path in line and ': ' in line:
                         threat_name = line.split(': ')[1].strip()
                         if threat_name.upper() != 'OK':
-                            print(f"🦠 Virus detected in {filename}: {threat_name}")
+                            self.logger.warning("Virus detected in %s: %s", filename, threat_name)
                             return False, threat_name
-                
+
                 # If we can't extract the specific threat name
-                print(f"🦠 Virus detected in {filename}: Unknown threat")
+                self.logger.warning("Virus detected in %s: Unknown threat", filename)
                 return False, "Unknown virus detected"
-            
+
             elif process.returncode == 0:
                 # File is clean
                 return True, None
-            
+
             else:
                 # Some other error occurred
-                print(f"⚠️ ClamAV scan error for {filename}: {stderr.strip()}")
+                self.logger.warning("ClamAV scan error for %s: %s", filename, stderr.strip())
                 return False, f"Scan error: {stderr.strip()}"
-                
-        except asyncio.TimeoutError:
-            print(f"⚠️ ClamAV scan timeout for {filename}")
+
+        except TimeoutError:
+            self.logger.warning("ClamAV scan timeout for %s", filename)
             return False, "Scan timeout"
         except Exception as e:
-            print(f"❌ Antivirus scan failed for {filename}: {e}")
+            self.logger.error("Antivirus scan failed for %s: %s", filename, e)
             return False, f"Scan error: {str(e)}"
         finally:
             # Clean up the temporary file
@@ -583,8 +588,8 @@ class AssetDocumentAgent:
                 Path(temp_path).unlink()
             except:
                 pass
-    
-    async def process_single_attachment(self, attachment_data: Dict[str, Any], email_data: Dict[str, Any]) -> ProcessingResult:
+
+    async def process_single_attachment(self, attachment_data: dict[str, Any], email_data: dict[str, Any]) -> ProcessingResult:
         """
         Process a single email attachment through Phase 1 pipeline.
         
@@ -597,60 +602,60 @@ class AssetDocumentAgent:
         """
         filename = attachment_data.get('filename', 'unknown_attachment')
         content = attachment_data.get('content', b'')
-        
-        print(f"  📄 Processing: {filename}")
-        
+
+        self.logger.info("Processing attachment %s", filename)
+
         try:
             # Step 1: File type validation
             if not self.validate_file_type(filename):
-                print(f"    ❌ Invalid file type: {Path(filename).suffix}")
+                self.logger.warning("Invalid file type for %s: %s", filename, Path(filename).suffix)
                 return ProcessingResult(
                     status=ProcessingStatus.INVALID_TYPE,
                     error_message=f"File type {Path(filename).suffix} not allowed"
                 )
-            
+
             # Step 2: File size validation
             file_size = len(content)
             if not self.validate_file_size(file_size):
-                print(f"    ❌ File too large: {file_size / (1024*1024):.1f} MB")
+                self.logger.warning("File too large for %s: %.1f MB", filename, file_size / (1024*1024))
                 return ProcessingResult(
                     status=ProcessingStatus.INVALID_TYPE,
                     error_message=f"File size {file_size / (1024*1024):.1f} MB exceeds limit"
                 )
-            
+
             # Step 3: Calculate SHA256 hash
             if not content:
                 return ProcessingResult(
                     status=ProcessingStatus.ERROR,
                     error_message="No file content available"
                 )
-            
+
             file_hash = self.calculate_file_hash(content)
-            print(f"    🔢 SHA256: {file_hash[:16]}...")
-            
+            self.logger.debug("SHA256(%s) = %s", filename, file_hash[:16])
+
             # Step 4: Check for duplicates (Phase 2)
             duplicate_id = await self.check_duplicate(file_hash)
             if duplicate_id:
-                print(f"    🔄 Duplicate detected (original: {duplicate_id})")
+                self.logger.info("Duplicate detected for %s (original: %s)", filename, duplicate_id)
                 return ProcessingResult(
                     status=ProcessingStatus.DUPLICATE,
                     file_hash=file_hash,
                     duplicate_of=duplicate_id
                 )
-            
+
             # Step 5: Antivirus scan
             is_clean, threat_name = await self.scan_file_antivirus(content, filename)
             if not is_clean:
-                print(f"    🦠 Quarantined: {threat_name}")
-                
+                self.logger.warning("File quarantined %s: %s", filename, threat_name)
+
                 return ProcessingResult(
                     status=ProcessingStatus.QUARANTINED,
                     file_hash=file_hash,
                     quarantine_reason=threat_name
                 )
-            
-            print(f"    ✅ File passed all Phase 1 validation checks")
-            
+
+            self.logger.info("File %s passed all Phase 1 validation checks", filename)
+
             # Prepare metadata for future phases
             metadata = {
                 'filename': filename,
@@ -659,29 +664,29 @@ class AssetDocumentAgent:
                 'sender_name': email_data.get('sender_name'),
                 'email_subject': email_data.get('subject'),
                 'email_date': email_data.get('date'),
-                'processing_date': datetime.now(timezone.utc).isoformat(),
+                'processing_date': datetime.now(UTC).isoformat(),
                 'file_extension': Path(filename).suffix.lower(),
                 'validation_passed': True
             }
-            
+
             return ProcessingResult(
                 status=ProcessingStatus.SUCCESS,
                 file_hash=file_hash,
                 confidence=1.0,  # Full confidence for validation
                 metadata=metadata
             )
-            
+
         except Exception as e:
-            print(f"    ❌ Processing error: {e}")
+            self.logger.error("Processing error for %s: %s", filename, e)
             return ProcessingResult(
                 status=ProcessingStatus.ERROR,
                 error_message=str(e)
             )
-    
-    def get_processing_stats(self) -> Dict[str, Any]:
+
+    def get_processing_stats(self) -> dict[str, Any]:
         """Get current processing statistics."""
         total = sum(self.stats.values())
-        
+
         return {
             'total_processed': total,
             'successful': self.stats['processed'],
@@ -690,15 +695,15 @@ class AssetDocumentAgent:
             'errors': self.stats['errors'],
             'success_rate': (self.stats['processed'] / total * 100) if total > 0 else 0.0
         }
-    
-    async def health_check(self) -> Dict[str, Any]:
+
+    async def health_check(self) -> dict[str, Any]:
         """Perform agent health check."""
         health = {
             'clamscan_path': self.clamscan_path is not None,
             'base_path_writable': False,
             'asset_configs_loaded': True
         }
-        
+
         try:
             # Check file system
             test_file = self.base_assets_path / '.health_check'
@@ -707,19 +712,19 @@ class AssetDocumentAgent:
             health['base_path_writable'] = True
         except:
             pass
-        
+
         return health
-    
+
     # ================================
     # PHASE 2: Asset & Sender Management
     # ================================
-    
+
     async def initialize_collections(self) -> bool:
         """Initialize Qdrant collections for asset management."""
         if not self.qdrant:
-            print("⚠️ Qdrant client not provided - skipping collection initialization")
+            self.logger.warning("Qdrant client not provided - skipping collection initialization")
             return False
-            
+
         try:
             # Assets collection with vector embeddings for semantic search
             if not await self._collection_exists(self.COLLECTIONS['assets']):
@@ -728,8 +733,8 @@ class AssetDocumentAgent:
                     vector_size=384,  # sentence-transformers dimension
                     description="Asset definitions with semantic embeddings"
                 )
-                print(f"✅ Created assets collection: {self.COLLECTIONS['assets']}")
-            
+                self.logger.info("Created assets collection: %s", self.COLLECTIONS['assets'])
+
             # Asset-Sender mappings (no vectors needed)
             if not await self._collection_exists(self.COLLECTIONS['asset_sender_mappings']):
                 await self._create_collection(
@@ -737,8 +742,8 @@ class AssetDocumentAgent:
                     vector_size=1,  # Minimal vector for Qdrant requirement
                     description="Many-to-many asset-sender relationships"
                 )
-                print(f"✅ Created asset-sender mappings collection")
-            
+                self.logger.info("Created asset-sender mappings collection")
+
             # Processed documents (no vectors needed)
             if not await self._collection_exists(self.COLLECTIONS['processed_documents']):
                 await self._create_collection(
@@ -746,8 +751,8 @@ class AssetDocumentAgent:
                     vector_size=1,  # Minimal vector
                     description="Processed document metadata and file hashes"
                 )
-                print(f"✅ Created processed documents collection")
-            
+                self.logger.info("Created processed documents collection")
+
             # Unknown senders (no vectors needed)
             if not await self._collection_exists(self.COLLECTIONS['unknown_senders']):
                 await self._create_collection(
@@ -755,15 +760,15 @@ class AssetDocumentAgent:
                     vector_size=1,  # Minimal vector
                     description="Unknown sender tracking and timeout management"
                 )
-                print(f"✅ Created unknown senders collection")
-            
-            print("🎯 All Qdrant collections initialized successfully")
+                self.logger.info("Created unknown senders collection")
+
+            self.logger.info("All Qdrant collections initialized successfully")
             return True
-            
+
         except Exception as e:
-            print(f"❌ Failed to initialize Qdrant collections: {e}")
+            self.logger.error("Failed to initialize Qdrant collections: %s", e)
             return False
-    
+
     async def _collection_exists(self, collection_name: str) -> bool:
         """Check if a Qdrant collection exists."""
         try:
@@ -771,7 +776,7 @@ class AssetDocumentAgent:
             return any(c.name == collection_name for c in collections.collections)
         except:
             return False
-    
+
     async def _create_collection(self, collection_name: str, vector_size: int, description: str) -> None:
         """Create a Qdrant collection."""
         self.qdrant.create_collection(
@@ -781,12 +786,12 @@ class AssetDocumentAgent:
                 distance=Distance.COSINE
             )
         )
-    
-    async def create_asset(self, 
+
+    async def create_asset(self,
                           deal_name: str,
                           asset_name: str,
                           asset_type: AssetType,
-                          identifiers: List[str] = None) -> str:
+                          identifiers: list[str] = None) -> str:
         """
         Create a new asset definition.
         
@@ -801,7 +806,7 @@ class AssetDocumentAgent:
         """
         deal_id = str(uuid.uuid4())
         folder_path = f"./assets/{deal_id}_{deal_name.replace(' ', '_')}"
-        
+
         asset = Asset(
             deal_id=deal_id,
             deal_name=deal_name,
@@ -809,15 +814,15 @@ class AssetDocumentAgent:
             asset_type=asset_type,
             folder_path=folder_path,
             identifiers=identifiers or [],
-            created_date=datetime.now(timezone.utc),
-            last_updated=datetime.now(timezone.utc)
+            created_date=datetime.now(UTC),
+            last_updated=datetime.now(UTC)
         )
-        
+
         if self.qdrant:
             try:
                 # Store asset in Qdrant (with dummy vector for now)
                 dummy_vector = [0.0] * 384  # Will be replaced with semantic embeddings
-                
+
                 point = PointStruct(
                     id=deal_id,
                     vector=dummy_vector,
@@ -832,27 +837,27 @@ class AssetDocumentAgent:
                         'last_updated': asset.last_updated.isoformat()
                     }
                 )
-                
+
                 self.qdrant.upsert(
                     collection_name=self.COLLECTIONS['assets'],
                     points=[point]
                 )
-                
-                print(f"✅ Created asset: {deal_name} ({deal_id})")
-                
+
+                self.logger.info("Created asset: %s (%s)", deal_name, deal_id)
+
             except Exception as e:
-                print(f"❌ Failed to store asset in Qdrant: {e}")
-                
+                self.logger.error("Failed to store asset in Qdrant: %s", e)
+
         # Create folder structure
         Path(folder_path).mkdir(parents=True, exist_ok=True)
-        
+
         return deal_id
-    
-    async def get_asset(self, deal_id: str) -> Optional[Asset]:
+
+    async def get_asset(self, deal_id: str) -> Asset | None:
         """Retrieve asset by deal_id."""
         if not self.qdrant:
             return None
-            
+
         try:
             search_result = self.qdrant.scroll(
                 collection_name=self.COLLECTIONS['assets'],
@@ -866,7 +871,7 @@ class AssetDocumentAgent:
                 ),
                 limit=1
             )
-            
+
             if search_result[0]:
                 payload = search_result[0][0].payload
                 return Asset(
@@ -879,26 +884,26 @@ class AssetDocumentAgent:
                     created_date=datetime.fromisoformat(payload['created_date']),
                     last_updated=datetime.fromisoformat(payload['last_updated'])
                 )
-                
+
             return None
-            
+
         except Exception as e:
-            print(f"❌ Failed to retrieve asset: {e}")
+            self.logger.error("Failed to retrieve asset: %s", e)
             return None
-    
-    async def list_assets(self) -> List[Asset]:
+
+    async def list_assets(self) -> list[Asset]:
         """List all assets."""
         assets = []
-        
+
         if not self.qdrant:
             return assets
-            
+
         try:
             search_result = self.qdrant.scroll(
                 collection_name=self.COLLECTIONS['assets'],
                 limit=100  # Adjust as needed
             )
-            
+
             for point in search_result[0]:
                 payload = point.payload
                 asset = Asset(
@@ -912,17 +917,17 @@ class AssetDocumentAgent:
                     last_updated=datetime.fromisoformat(payload['last_updated'])
                 )
                 assets.append(asset)
-                
+
         except Exception as e:
-            print(f"❌ Failed to list assets: {e}")
-            
+            self.logger.error("Failed to list assets: %s", e)
+
         return assets
-    
+
     async def create_asset_sender_mapping(self,
                                         asset_id: str,
                                         sender_email: str,
                                         confidence: float,
-                                        document_types: List[str] = None) -> str:
+                                        document_types: list[str] = None) -> str:
         """
         Create asset-sender mapping.
         
@@ -936,22 +941,22 @@ class AssetDocumentAgent:
             Mapping ID (UUID)
         """
         mapping_id = str(uuid.uuid4())
-        
+
         mapping = AssetSenderMapping(
             mapping_id=mapping_id,
             asset_id=asset_id,
             sender_email=sender_email.lower(),
             confidence=confidence,
             document_types=document_types or [],
-            created_date=datetime.now(timezone.utc),
-            last_activity=datetime.now(timezone.utc)
+            created_date=datetime.now(UTC),
+            last_activity=datetime.now(UTC)
         )
-        
+
         if self.qdrant:
             try:
                 # Store mapping in Qdrant
                 dummy_vector = [0.0]  # Minimal vector
-                
+
                 point = PointStruct(
                     id=mapping_id,
                     vector=dummy_vector,
@@ -966,26 +971,26 @@ class AssetDocumentAgent:
                         'email_count': 0
                     }
                 )
-                
+
                 self.qdrant.upsert(
                     collection_name=self.COLLECTIONS['asset_sender_mappings'],
                     points=[point]
                 )
-                
-                print(f"✅ Created asset-sender mapping: {sender_email} → {asset_id}")
-                
+
+                self.logger.info("Created asset-sender mapping: %s → %s", sender_email, asset_id)
+
             except Exception as e:
-                print(f"❌ Failed to store mapping in Qdrant: {e}")
-                
+                self.logger.error("Failed to store mapping in Qdrant: %s", e)
+
         return mapping_id
-    
-    async def get_sender_assets(self, sender_email: str) -> List[Dict[str, Any]]:
+
+    async def get_sender_assets(self, sender_email: str) -> list[dict[str, Any]]:
         """Get all assets associated with a sender."""
         assets = []
-        
+
         if not self.qdrant:
             return assets
-            
+
         try:
             search_result = self.qdrant.scroll(
                 collection_name=self.COLLECTIONS['asset_sender_mappings'],
@@ -999,7 +1004,7 @@ class AssetDocumentAgent:
                 ),
                 limit=50
             )
-            
+
             for point in search_result[0]:
                 payload = point.payload
                 assets.append({
@@ -1009,25 +1014,25 @@ class AssetDocumentAgent:
                     'document_types': payload['document_types'],
                     'email_count': payload.get('email_count', 0)
                 })
-                
+
         except Exception as e:
-            print(f"❌ Failed to get sender assets: {e}")
-            
+            self.logger.error("Failed to get sender assets: %s", e)
+
         return assets
-    
-    async def list_asset_sender_mappings(self) -> List[Dict[str, Any]]:
+
+    async def list_asset_sender_mappings(self) -> list[dict[str, Any]]:
         """List all asset-sender mappings."""
         mappings = []
-        
+
         if not self.qdrant:
             return mappings
-            
+
         try:
             search_result = self.qdrant.scroll(
                 collection_name=self.COLLECTIONS['asset_sender_mappings'],
                 limit=100
             )
-            
+
             for point in search_result[0]:
                 payload = point.payload
                 mappings.append({
@@ -1040,20 +1045,20 @@ class AssetDocumentAgent:
                     'last_activity': payload['last_activity'],
                     'email_count': payload.get('email_count', 0)
                 })
-                
+
         except Exception as e:
-            print(f"❌ Failed to list sender mappings: {e}")
-            
+            self.logger.error("Failed to list sender mappings: %s", e)
+
         return mappings
-    
+
     async def delete_asset_sender_mapping(self, mapping_id: str) -> bool:
         """Delete an asset-sender mapping by mapping_id."""
         if not self.qdrant:
-            print(f"❌ Qdrant client not available for deletion")
+            self.logger.error("Qdrant client not available for deletion")
             return False
-            
-        print(f"🗑️ Attempting to delete sender mapping: {mapping_id}")
-        
+
+        self.logger.info("Attempting to delete sender mapping: %s", mapping_id)
+
         try:
             # First verify the mapping exists
             search_result = self.qdrant.scroll(
@@ -1068,13 +1073,13 @@ class AssetDocumentAgent:
                 ),
                 limit=1
             )
-            
+
             if not search_result[0]:
-                print(f"❌ Sender mapping not found in database: {mapping_id}")
+                self.logger.error("Sender mapping not found in database: %s", mapping_id)
                 return False
-                
-            print(f"✓ Found mapping to delete: {mapping_id}")
-            
+
+            self.logger.debug("Found mapping to delete: %s", mapping_id)
+
             # Delete by point ID (which should be the mapping_id)
             delete_result = self.qdrant.delete(
                 collection_name=self.COLLECTIONS['asset_sender_mappings'],
@@ -1082,10 +1087,10 @@ class AssetDocumentAgent:
                     points=[mapping_id]
                 )
             )
-            
-            print(f"✅ Delete operation completed for mapping: {mapping_id}")
-            print(f"✅ Delete result: {delete_result}")
-            
+
+            self.logger.info("Delete operation completed for mapping: %s", mapping_id)
+            self.logger.debug("Delete result: %s", delete_result)
+
             # Verify deletion
             verify_result = self.qdrant.scroll(
                 collection_name=self.COLLECTIONS['asset_sender_mappings'],
@@ -1099,21 +1104,21 @@ class AssetDocumentAgent:
                 ),
                 limit=1
             )
-            
+
             if not verify_result[0]:
-                print(f"✅ Confirmed: Sender mapping successfully deleted: {mapping_id}")
+                self.logger.info("Confirmed: Sender mapping successfully deleted: %s", mapping_id)
                 return True
             else:
-                print(f"❌ Deletion verification failed - mapping still exists: {mapping_id}")
+                self.logger.error("Deletion verification failed - mapping still exists: %s", mapping_id)
                 return False
-            
+
         except Exception as e:
-            print(f"❌ Failed to delete sender mapping {mapping_id}: {e}")
+            self.logger.error("Failed to delete sender mapping %s: %s", mapping_id, e)
             import traceback
-            print(f"❌ Full error traceback: {traceback.format_exc()}")
+            self.logger.error("Full error traceback: %s", traceback.format_exc())
             return False
-    
-    async def check_duplicate(self, file_hash: str) -> Optional[str]:
+
+    async def check_duplicate(self, file_hash: str) -> str | None:
         """
         Check if file hash already exists in processed documents.
         
@@ -1122,7 +1127,7 @@ class AssetDocumentAgent:
         """
         if not self.qdrant:
             return None
-            
+
         try:
             search_result = self.qdrant.scroll(
                 collection_name=self.COLLECTIONS['processed_documents'],
@@ -1136,64 +1141,74 @@ class AssetDocumentAgent:
                 ),
                 limit=1
             )
-            
+
             if search_result[0]:  # Found duplicate
                 return search_result[0][0].payload.get('document_id')
-            
+
             return None
-            
+
         except Exception as e:
-            print(f"⚠️ Error checking for duplicates: {e}")
+            self.logger.warning("Error checking for duplicates: %s", e)
             return None
-    
-    async def store_processed_document(self, 
+
+    async def store_processed_document(self,
                                      file_hash: str,
                                      processing_result: ProcessingResult,
                                      asset_id: str = None) -> str:
         """Store processed document metadata."""
         document_id = str(uuid.uuid4())
-        
+
         if not self.qdrant:
             return document_id
-            
+
         try:
             dummy_vector = [0.0]  # Minimal vector
-            
+
             payload = {
                 'document_id': document_id,
                 'file_hash': file_hash,
                 'status': processing_result.status.value,
                 'confidence': processing_result.confidence,
-                'processing_date': datetime.now(timezone.utc).isoformat(),
+                'processing_date': datetime.now(UTC).isoformat(),
                 'metadata': processing_result.metadata or {}
             }
-            
+
             if asset_id:
                 payload['asset_id'] = asset_id
-                
+
             if processing_result.file_path:
                 payload['file_path'] = str(processing_result.file_path)
-                
+
             point = PointStruct(
                 id=document_id,
                 vector=dummy_vector,
                 payload=payload
             )
-            
+
             self.qdrant.upsert(
                 collection_name=self.COLLECTIONS['processed_documents'],
                 points=[point]
             )
-            
+
+            self.logger.info("💾 Saved to: %s", processing_result.file_path)
+
+            # Store document metadata in Qdrant
+            if self.qdrant and file_hash:
+                document_id = await self.store_processed_document(
+                    file_hash, processing_result, asset_id
+                )
+                self.logger.info("📊 Document metadata stored: %s...", document_id[:8])
+
+            return document_id
+
         except Exception as e:
-            print(f"❌ Failed to store processed document: {e}")
-            
-        return document_id
-    
+            self.logger.error("Failed to store processed document: %s", e)
+            return None
+
     # ================================
     # PHASE 3: Document Classification
     # ================================
-    
+
     # Document classification patterns by asset type
     CLASSIFICATION_PATTERNS = {
         AssetType.COMMERCIAL_REAL_ESTATE: {
@@ -1283,7 +1298,7 @@ class AssetDocumentAgent:
             ]
         }
     }
-    
+
     # Asset name patterns for fuzzy matching
     ASSET_KEYWORDS = {
         'commercial_real_estate': [
@@ -1303,12 +1318,12 @@ class AssetDocumentAgent:
             'power', 'water', 'gas', 'pipeline', 'renewable', 'solar', 'wind'
         ]
     }
-    
-    def classify_document(self, 
-                         filename: str, 
+
+    def classify_document(self,
+                         filename: str,
                          email_subject: str,
                          email_body: str = "",
-                         asset_type: AssetType = None) -> Tuple[DocumentCategory, float]:
+                         asset_type: AssetType = None) -> tuple[DocumentCategory, float]:
         """
         Classify document based on filename, email subject, and content.
         
@@ -1323,21 +1338,21 @@ class AssetDocumentAgent:
         """
         # Combine all text for analysis
         combined_text = f"{filename} {email_subject} {email_body}".lower()
-        
+
         # If asset type is known, only check patterns for that type
         if asset_type and asset_type in self.CLASSIFICATION_PATTERNS:
             patterns_to_check = {asset_type: self.CLASSIFICATION_PATTERNS[asset_type]}
         else:
             patterns_to_check = self.CLASSIFICATION_PATTERNS
-        
+
         best_category = DocumentCategory.UNKNOWN
         best_score = 0.0
-        
+
         for asset_type_key, categories in patterns_to_check.items():
             for category, patterns in categories.items():
                 score = 0.0
                 matches = 0
-                
+
                 for pattern in patterns:
                     if re.search(pattern, combined_text):
                         matches += 1
@@ -1348,26 +1363,26 @@ class AssetDocumentAgent:
                             score += 0.3
                         else:
                             score += 0.1
-                
+
                 # Normalize score by number of patterns
                 if patterns:
                     normalized_score = min(score / len(patterns), 1.0)
-                    
+
                     # Boost score if multiple patterns match
                     if matches > 1:
                         normalized_score = min(normalized_score * 1.2, 1.0)
-                    
+
                     if normalized_score > best_score:
                         best_score = normalized_score
                         best_category = category
-        
+
         return best_category, best_score
-    
-    def identify_asset_from_content(self, 
+
+    def identify_asset_from_content(self,
                                    email_subject: str,
                                    email_body: str = "",
                                    filename: str = "",
-                                   known_assets: List[Asset] = None) -> List[Tuple[str, float]]:
+                                   known_assets: list[Asset] = None) -> list[tuple[str, float]]:
         """
         Identify potential assets mentioned in email content using fuzzy matching.
         
@@ -1382,68 +1397,68 @@ class AssetDocumentAgent:
         """
         if not known_assets or not levenshtein_distance:
             return []
-        
+
         # Combine text for analysis
         combined_text = f"{email_subject} {email_body} {filename}".lower()
-        
+
         asset_matches = []
-        
+
         for asset in known_assets:
             max_confidence = 0.0
-            
+
             # Check all asset identifiers
             all_identifiers = [asset.deal_name, asset.asset_name] + asset.identifiers
-            
+
             for identifier in all_identifiers:
                 if not identifier:
                     continue
-                    
+
                 identifier_lower = identifier.lower()
-                
+
                 # Exact match (highest confidence)
                 if identifier_lower in combined_text:
                     confidence = 0.95
                     max_confidence = max(max_confidence, confidence)
                     continue
-                
+
                 # Fuzzy matching with Levenshtein distance
                 words = combined_text.split()
                 for word_group_size in [3, 2, 1]:  # Check phrases of different lengths
                     for i in range(len(words) - word_group_size + 1):
                         phrase = " ".join(words[i:i + word_group_size])
-                        
+
                         if len(phrase) < 3:  # Skip very short phrases
                             continue
-                            
+
                         # Calculate similarity
                         distance = levenshtein_distance(identifier_lower, phrase)
                         max_len = max(len(identifier_lower), len(phrase))
-                        
+
                         if max_len > 0:
                             similarity = 1 - (distance / max_len)
-                            
+
                             # Threshold for fuzzy matches
                             if similarity >= 0.8:
                                 confidence = similarity * 0.9  # Slightly lower than exact match
                                 max_confidence = max(max_confidence, confidence)
-            
+
             # Add partial word matching for specific keywords
             asset_type_keywords = self.ASSET_KEYWORDS.get(asset.asset_type.value, [])
             keyword_matches = sum(1 for keyword in asset_type_keywords if keyword in combined_text)
-            
+
             if keyword_matches > 0:
                 keyword_boost = min(keyword_matches * 0.1, 0.3)
                 max_confidence = min(max_confidence + keyword_boost, 1.0)
-            
+
             if max_confidence > 0.5:  # Only include reasonable matches
                 asset_matches.append((asset.deal_id, max_confidence))
-        
+
         # Sort by confidence (highest first)
         asset_matches.sort(key=lambda x: x[1], reverse=True)
-        
+
         return asset_matches
-    
-    def determine_confidence_level(self, 
+
+    def determine_confidence_level(self,
                                   document_confidence: float,
                                   asset_confidence: float,
                                   sender_known: bool) -> ConfidenceLevel:
@@ -1464,15 +1479,15 @@ class AssetDocumentAgent:
             'asset': 0.4,
             'sender': 0.2
         }
-        
+
         sender_score = 1.0 if sender_known else 0.0
-        
+
         overall_confidence = (
             document_confidence * weights['document'] +
             asset_confidence * weights['asset'] +
             sender_score * weights['sender']
         )
-        
+
         # Map to confidence levels with thresholds from requirements
         if overall_confidence >= 0.90:
             return ConfidenceLevel.HIGH      # Auto-process
@@ -1482,10 +1497,10 @@ class AssetDocumentAgent:
             return ConfidenceLevel.LOW       # Save uncategorized
         else:
             return ConfidenceLevel.VERY_LOW  # Human review required
-    
-    async def enhanced_process_attachment(self, 
-                                        attachment_data: Dict[str, Any], 
-                                        email_data: Dict[str, Any]) -> ProcessingResult:
+
+    async def enhanced_process_attachment(self,
+                                        attachment_data: dict[str, Any],
+                                        email_data: dict[str, Any]) -> ProcessingResult:
         """
         Enhanced attachment processing with Phase 3 classification and intelligence.
         
@@ -1498,72 +1513,72 @@ class AssetDocumentAgent:
         """
         # Start with Phase 1 & 2 processing
         result = await self.process_single_attachment(attachment_data, email_data)
-        
+
         if result.status != ProcessingStatus.SUCCESS:
             return result
-        
+
         # Phase 3: Add document classification and asset identification
         filename = attachment_data.get('filename', '')
         email_subject = email_data.get('subject', '')
         email_body = email_data.get('body', '')
         sender_email = email_data.get('sender_email', '')
-        
-        print(f"    🧠 Phase 3: Analyzing content and identifying assets...")
-        
+
+        self.logger.info("🧠 Phase 3: Analyzing content and identifying assets...")
+
         # Step 1: Document Classification
         document_category, doc_confidence = self.classify_document(
             filename, email_subject, email_body
         )
-        
-        print(f"    📋 Document classified as: {document_category.value} (confidence: {doc_confidence:.2f})")
-        
+
+        self.logger.info("📋 Document classified as: %s (confidence: %.2f)", document_category.value, doc_confidence)
+
         # Step 2: Asset Identification
         known_assets = await self.list_assets()
         asset_matches = self.identify_asset_from_content(
             email_subject, email_body, filename, known_assets
         )
-        
+
         matched_asset_id = None
         asset_confidence = 0.0
-        
+
         if asset_matches:
             matched_asset_id, asset_confidence = asset_matches[0]  # Best match
             asset = await self.get_asset(matched_asset_id)
             asset_name = asset.deal_name if asset else matched_asset_id[:8]
-            print(f"    🎯 Best asset match: {asset_name} (confidence: {asset_confidence:.2f})")
+            self.logger.info("🎯 Best asset match: %s (confidence: %.2f)", asset_name, asset_confidence)
         else:
-            print(f"    ❓ No asset matches found")
-        
+            self.logger.info("❓ No asset matches found")
+
         # Step 3: Check if sender is known
         sender_assets = await self.get_sender_assets(sender_email)
         sender_known = len(sender_assets) > 0
-        
+
         if sender_known:
-            print(f"    👤 Known sender: {len(sender_assets)} asset(s) associated")
+            self.logger.info("👤 Known sender: %d asset(s) associated", len(sender_assets))
             # If sender has assets but we didn't match content, use sender's highest confidence asset
             if not matched_asset_id and sender_assets:
                 best_sender_asset = max(sender_assets, key=lambda x: x['confidence'])
                 matched_asset_id = best_sender_asset['asset_id']
                 asset_confidence = best_sender_asset['confidence'] * 0.8  # Slightly reduce confidence
-                print(f"    🔗 Using sender's primary asset: {matched_asset_id[:8]}...")
+                self.logger.info("🔗 Using sender's primary asset: %s...", matched_asset_id[:8])
         else:
-            print(f"    ❓ Unknown sender: {sender_email}")
-        
+            self.logger.info("❓ Unknown sender: %s", sender_email)
+
         # Step 4: Determine overall confidence level
         confidence_level = self.determine_confidence_level(
             doc_confidence, asset_confidence, sender_known
         )
-        
+
         confidence_icons = {
             ConfidenceLevel.HIGH: "🟢",
-            ConfidenceLevel.MEDIUM: "🟡", 
+            ConfidenceLevel.MEDIUM: "🟡",
             ConfidenceLevel.LOW: "🟠",
             ConfidenceLevel.VERY_LOW: "🔴"
         }
-        
+
         icon = confidence_icons[confidence_level]
-        print(f"    {icon} Overall confidence: {confidence_level.value}")
-        
+        self.logger.info("%s Overall confidence: %s", icon, confidence_level.value)
+
         # Step 5: Update result with Phase 3 data
         result.document_category = document_category
         result.confidence_level = confidence_level
@@ -1576,14 +1591,14 @@ class AssetDocumentAgent:
             'asset_matches': asset_matches[:3],  # Top 3 matches
             'sender_asset_count': len(sender_assets)
         }
-        
+
         return result
 
-    async def save_attachment_to_asset_folder(self, 
+    async def save_attachment_to_asset_folder(self,
                                            attachment_content: bytes,
                                            filename: str,
                                            processing_result: ProcessingResult,
-                                           asset_id: Optional[str] = None) -> Optional[Path]:
+                                           asset_id: str | None = None) -> Path | None:
         """
         Save processed attachment to the appropriate asset folder.
         
@@ -1603,7 +1618,7 @@ class AssetDocumentAgent:
                 asset = await self.get_asset(asset_id)
                 if asset:
                     base_folder = Path(asset.folder_path)
-                    
+
                     # Create document category subfolder
                     if processing_result.document_category and processing_result.document_category != DocumentCategory.UNKNOWN:
                         category_folder = base_folder / processing_result.document_category.value
@@ -1616,15 +1631,15 @@ class AssetDocumentAgent:
             else:
                 # Low confidence or no asset match - save to review folder
                 base_folder = self.base_assets_path / "to_be_reviewed"
-                
+
                 if processing_result.confidence_level == ConfidenceLevel.LOW:
                     category_folder = base_folder / "low_confidence"
                 else:
                     category_folder = base_folder / "very_low_confidence"
-            
+
             # Create folder structure
             category_folder.mkdir(parents=True, exist_ok=True)
-            
+
             # Generate unique filename to avoid conflicts
             file_path = category_folder / filename
             counter = 1
@@ -1635,25 +1650,25 @@ class AssetDocumentAgent:
                 new_filename = f"{stem}_{counter:03d}{suffix}"
                 file_path = category_folder / new_filename
                 counter += 1
-            
+
             # Save the file
             with open(file_path, 'wb') as f:
                 f.write(attachment_content)
-            
-            self.logger.info(f"File saved to: {file_path}")
-            
+
+            self.logger.info(f"💾 Saved to: {file_path}")
+
             # Update processing result with file path
             processing_result.file_path = file_path
-            
+
             return file_path
-            
+
         except Exception as e:
             self.logger.error(f"Failed to save attachment {filename}: {e}")
             return None
 
-    async def process_and_save_attachment(self, 
-                                        attachment_data: Dict[str, Any], 
-                                        email_data: Dict[str, Any],
+    async def process_and_save_attachment(self,
+                                        attachment_data: dict[str, Any],
+                                        email_data: dict[str, Any],
                                         save_to_disk: bool = True) -> ProcessingResult:
         """
         Complete attachment processing pipeline including file saving.
@@ -1671,32 +1686,32 @@ class AssetDocumentAgent:
         """
         # Step 1: Enhanced processing (validation + classification)
         result = await self.enhanced_process_attachment(attachment_data, email_data)
-        
+
         # Step 2: Save to disk if successful and requested
         if save_to_disk and result.status == ProcessingStatus.SUCCESS:
             content = attachment_data.get('content', b'')
             filename = attachment_data.get('filename', 'unknown_attachment')
-            
+
             if content:
                 file_path = await self.save_attachment_to_asset_folder(
                     content, filename, result, result.matched_asset_id
                 )
-                
+
                 if file_path:
-                    print(f"    💾 Saved to: {file_path}")
-                    
+                    self.logger.info("💾 Saved to: %s", file_path)
+
                     # Store document metadata in Qdrant
                     if self.qdrant and result.file_hash:
                         document_id = await self.store_processed_document(
                             result.file_hash, result, result.matched_asset_id
                         )
-                        print(f"    📊 Document metadata stored: {document_id[:8]}...")
+                        self.logger.info("📊 Document metadata stored: %s...", document_id[:8])
                 else:
-                    print(f"    ❌ Failed to save file to disk")
-        
+                    self.logger.warning("❌ Failed to save file to disk")
+
         return result
 
-    async def process_attachments_parallel(self, email_attachments: List[Dict[str, Any]], email_data: Dict[str, Any]) -> List[ProcessingResult]:
+    async def process_attachments_parallel(self, email_attachments: list[dict[str, Any]], email_data: dict[str, Any]) -> list[ProcessingResult]:
         """
         Process multiple email attachments in parallel for maximum performance.
         
@@ -1714,16 +1729,16 @@ class AssetDocumentAgent:
         """
         if not email_attachments:
             return []
-        
+
         start_time = asyncio.get_event_loop().time()
         self.logger.info(f"🔄 Starting parallel processing of {len(email_attachments)} attachments")
-        
+
         # Create tasks for parallel processing
         tasks = []
         for i, attachment_data in enumerate(email_attachments):
             filename = attachment_data.get('filename', f'attachment_{i}')
             self.logger.debug(f"Creating task for {filename}")
-            
+
             # Create coroutine for processing this attachment
             task = self.process_and_save_attachment(
                 attachment_data=attachment_data,
@@ -1731,18 +1746,18 @@ class AssetDocumentAgent:
                 save_to_disk=True
             )
             tasks.append(task)
-        
+
         try:
             # Execute all tasks concurrently with proper error handling
             results = await asyncio.gather(*tasks, return_exceptions=True)
-            
+
             # Convert exceptions to error ProcessingResults
             final_results = []
             for i, result in enumerate(results):
                 if isinstance(result, Exception):
                     filename = email_attachments[i].get('filename', f'attachment_{i}')
                     self.logger.error(f"Error processing {filename}: {result}")
-                    
+
                     error_result = ProcessingResult(
                         status=ProcessingStatus.ERROR,
                         error_message=str(result)
@@ -1750,26 +1765,26 @@ class AssetDocumentAgent:
                     final_results.append(error_result)
                 else:
                     final_results.append(result)
-            
+
             end_time = asyncio.get_event_loop().time()
             total_time = end_time - start_time
-            
+
             # Log performance improvement
             sequential_estimate = len(email_attachments) * 10  # Assume 10s per attachment sequentially
             improvement = ((sequential_estimate - total_time) / sequential_estimate) * 100
-            
+
             self.logger.info(f"✅ Parallel processing completed in {total_time:.1f}s")
             self.logger.info(f"🚀 Estimated performance improvement: {improvement:.1f}% faster than sequential")
             self.logger.info(f"   Sequential estimate: {sequential_estimate:.1f}s vs Parallel actual: {total_time:.1f}s")
-            
+
             # Update statistics
             self.processing_times.append(total_time)
-            
+
             return final_results
-            
+
         except Exception as e:
             self.logger.error(f"Critical error in parallel processing: {e}")
-            
+
             # Fallback: return error results for all attachments
             return [
                 ProcessingResult(
@@ -1779,7 +1794,7 @@ class AssetDocumentAgent:
                 for _ in email_attachments
             ]
 
-    async def scan_files_antivirus_parallel(self, file_data_list: List[Tuple[bytes, str]]) -> List[Tuple[bool, Optional[str]]]:
+    async def scan_files_antivirus_parallel(self, file_data_list: list[tuple[bytes, str]]) -> list[tuple[bool, str | None]]:
         """
         Scan multiple files with ClamAV in parallel for maximum performance.
         
@@ -1794,20 +1809,20 @@ class AssetDocumentAgent:
         """
         if not file_data_list:
             return []
-        
+
         start_time = asyncio.get_event_loop().time()
         self.logger.info(f"🦠 Starting parallel virus scanning of {len(file_data_list)} files")
-        
+
         # Create tasks for parallel scanning
         tasks = []
         for file_content, filename in file_data_list:
             task = self.scan_file_antivirus(file_content, filename)
             tasks.append(task)
-        
+
         try:
             # Execute all scans concurrently
             results = await asyncio.gather(*tasks, return_exceptions=True)
-            
+
             # Convert exceptions to failed scan results
             final_results = []
             for i, result in enumerate(results):
@@ -1817,22 +1832,22 @@ class AssetDocumentAgent:
                     final_results.append((False, f"Scan error: {str(result)}"))
                 else:
                     final_results.append(result)
-            
+
             end_time = asyncio.get_event_loop().time()
             total_time = end_time - start_time
-            
+
             # Log performance improvement
             sequential_estimate = len(file_data_list) * 10  # Assume 10s per scan sequentially
             improvement = ((sequential_estimate - total_time) / sequential_estimate) * 100
-            
+
             self.logger.info(f"✅ Parallel virus scanning completed in {total_time:.1f}s")
             self.logger.info(f"🚀 Estimated performance improvement: {improvement:.1f}% faster than sequential")
-            
+
             return final_results
-            
+
         except Exception as e:
             self.logger.error(f"Critical error in parallel virus scanning: {e}")
-            
+
             # Fallback: return scan error for all files
             return [
                 (False, f"Parallel scan failed: {str(e)}")
@@ -1842,29 +1857,29 @@ class AssetDocumentAgent:
 # Example usage and testing
 async def test_asset_document_agent() -> None:
     """Test function for the Asset Document Agent Phase 1 implementation."""
-    print("🧪 Testing Asset Document Agent - Phase 1")
-    print("=" * 60)
-    
+    print("🧪 Testing Asset Document Agent - Phase 1")  # noqa: T201
+    print("=" * 60)  # noqa: T201
+
     # Initialize agent without Qdrant for basic testing
     agent = AssetDocumentAgent()
-    
+
     # Test health check
     health = await agent.health_check()
-    print(f"🏥 Agent health check: {health}")
-    
+    print(f"🏥 Agent health check: {health}")  # noqa: T201
+
     # Test file validation
-    print(f"\n📋 Testing file validation:")
-    print(f"   PDF file: {agent.validate_file_type('document.pdf')}")
-    print(f"   Excel file: {agent.validate_file_type('spreadsheet.xlsx')}")
-    print(f"   Executable: {agent.validate_file_type('malware.exe')}")
-    
+    print("\n📋 Testing file validation:")  # noqa: T201
+    print(f"   PDF file: {agent.validate_file_type('document.pdf')}")  # noqa: T201
+    print(f"   Excel file: {agent.validate_file_type('spreadsheet.xlsx')}")  # noqa: T201
+    print(f"   Executable: {agent.validate_file_type('malware.exe')}")  # noqa: T201
+
     # Test file hash calculation
     test_content = b"This is test file content"
     file_hash = agent.calculate_file_hash(test_content)
-    print(f"\n🔢 Test file hash: {file_hash[:16]}...")
-    
+    print(f"\n🔢 Test file hash: {file_hash[:16]}...")  # noqa: T201
+
     # Test attachment processing
-    print(f"\n📎 Testing attachment processing:")
+    print("\n📎 Testing attachment processing:")  # noqa: T201
     test_attachment = {
         'filename': 'test_document.pdf',
         'content': test_content
@@ -1875,22 +1890,22 @@ async def test_asset_document_agent() -> None:
         'subject': 'Test Email',
         'date': datetime.now().isoformat()
     }
-    
+
     result = await agent.process_single_attachment(test_attachment, test_email)
-    print(f"   Result: {result.status.value}")
-    print(f"   Hash: {result.file_hash[:16] if result.file_hash else 'None'}...")
-    print(f"   Confidence: {result.confidence}")
-    
+    print(f"   Result: {result.status.value}")  # noqa: T201
+    print(f"   Hash: {result.file_hash[:16] if result.file_hash else 'None'}...")  # noqa: T201
+    print(f"   Confidence: {result.confidence}")  # noqa: T201
+
     # Get statistics
     stats = agent.get_processing_stats()
-    print(f"\n📊 Processing statistics: {stats}")
-    
-    print("\n📋 Phase 1 implementation complete!")
-    print("   ✅ File type and size validation")
-    print("   ✅ SHA256 hashing")
-    print("   ✅ ClamAV antivirus integration")
-    print("   ✅ Processing pipeline foundation")
-    print("   ✅ Health monitoring")
+    print(f"\n📊 Processing statistics: {stats}")  # noqa: T201
+
+    print("\n📋 Phase 1 implementation complete!")  # noqa: T201
+    print("   ✅ File type and size validation")  # noqa: T201
+    print("   ✅ SHA256 hashing")  # noqa: T201
+    print("   ✅ ClamAV antivirus integration")  # noqa: T201
+    print("   ✅ Processing pipeline foundation")  # noqa: T201
+    print("   ✅ Health monitoring")  # noqa: T201
 
 if __name__ == "__main__":
-    asyncio.run(test_asset_document_agent()) 
+    asyncio.run(test_asset_document_agent())
