@@ -43,29 +43,29 @@ Copyright 2025 by Inveniam Capital Partners, LLC and Rick Bunker
 """
 
 import asyncio
+import os
 import re
-from typing import Dict, List, Optional, Any, Set, Tuple, Union
-from dataclasses import dataclass, field
+
+# Logging system integration
+import sys
+from dataclasses import dataclass
+from datetime import UTC, datetime
 from enum import Enum
-from datetime import datetime, UTC
-import hashlib
-import json
+from typing import Any
+
+from ..memory.contact import ContactMemory
+from ..memory.episodic import EpisodicMemory
 
 # Memory system integration
 from ..memory.procedural import ProceduralMemory
 from ..memory.semantic import SemanticMemory
-from ..memory.episodic import EpisodicMemory
-from ..memory.contact import ContactMemory, ContactType as MemContactType, ContactConfidence as MemContactConfidence
 
 # Email message structure
 from .supervisor import EmailMessage
 
-# Logging system integration
-import sys
-import os
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 
-from utils.logging_system import get_logger, log_function, log_debug, log_info
+from utils.logging_system import get_logger, log_function
 
 # Initialize logger
 logger = get_logger(__name__)
@@ -123,29 +123,29 @@ class ContactInfo:
         metadata: Additional structured data and metrics
     """
     email: str
-    name: Optional[str] = None
-    first_name: Optional[str] = None
-    last_name: Optional[str] = None
-    phone: Optional[str] = None
-    organization: Optional[str] = None
-    title: Optional[str] = None
-    address: Optional[str] = None
+    name: str | None = None
+    first_name: str | None = None
+    last_name: str | None = None
+    phone: str | None = None
+    organization: str | None = None
+    title: str | None = None
+    address: str | None = None
     contact_type: ContactType = ContactType.UNKNOWN
     confidence: ContactConfidence = ContactConfidence.LOW
-    source_email_id: Optional[str] = None
+    source_email_id: str | None = None
     extraction_reasoning: str = ""
-    metadata: Optional[Dict[str, Any]] = None
-    
+    metadata: dict[str, Any] | None = None
+
     def __post_init__(self) -> None:
         """Initialize metadata dictionary if not provided."""
         if self.metadata is None:
             self.metadata = {}
-        
+
         # Validate email format
         if not self.email or '@' not in self.email:
             raise ValueError(f"Invalid email address: {self.email}")
-    
-    def to_dict(self) -> Dict[str, Any]:
+
+    def to_dict(self) -> dict[str, Any]:
         """Convert contact info to dictionary for storage."""
         return {
             'email': self.email,
@@ -162,9 +162,9 @@ class ContactInfo:
             'extraction_reasoning': self.extraction_reasoning,
             'metadata': self.metadata or {}
         }
-    
+
     @classmethod
-    def from_dict(cls, data: Dict[str, Any]) -> 'ContactInfo':
+    def from_dict(cls, data: dict[str, Any]) -> 'ContactInfo':
         """Create ContactInfo from dictionary data."""
         return cls(
             email=data['email'],
@@ -202,7 +202,7 @@ class ContactExtractor:
         episodic_memory: Stores extraction history and feedback
         contact_memory: Stores extracted contact information
     """
-    
+
     def __init__(self) -> None:
         """
         Initialize the contact extraction agent.
@@ -215,11 +215,11 @@ class ContactExtractor:
         self.semantic_memory = SemanticMemory(max_items=1000)
         self.episodic_memory = EpisodicMemory(max_items=1000)
         self.contact_memory = ContactMemory(max_items=5000)
-        
+
         # Initialize logger
         self.logger = get_logger(f"{__name__}.{self.__class__.__name__}")
         self.logger.info("Initializing Contact Extraction Agent")
-        
+
         # Patterns for identifying automated/bulk emails
         self.no_reply_patterns = [
             r'no[-_]?reply',
@@ -237,7 +237,7 @@ class ContactExtractor:
             r'robot',
             r'bounce'
         ]
-        
+
         # Common bulk/marketing domains to exclude
         self.bulk_domains = {
             'mailchimp.com', 'constantcontact.com', 'sendgrid.net',
@@ -247,30 +247,30 @@ class ContactExtractor:
             'madmimi.com', 'verticalresponse.com', 'icontact.com',
             'benchmarkemail.com', 'emailbrain.com', 'silverpop.com'
         }
-        
+
         # Patterns for extracting contact information
         self.phone_pattern = re.compile(
             r'\b(?:\+?1[-.\s]?)?\(?([0-9]{3})\)?[-.\s]?([0-9]{3})[-.\s]?([0-9]{4})\b'
         )
         self.name_pattern = re.compile(r'\b([A-Z][a-z]+(?:\s+[A-Z][a-z]+)*)\b')
-        
+
         # Patterns for identifying personal vs automated content
         self.personal_indicators = [
             r'\bthanks?\b', r'\bthank you\b', r'\bregards\b', r'\bbest\b',
             r'\bsincerely\b', r'\bcheers\b', r'\bhope\b', r'\bwish\b',
             r'\bfyi\b', r'\bbtw\b', r'\blmk\b', r'\basap\b'
         ]
-        
+
         self.automated_indicators = [
             r'\bautomatic\b', r'\bgenerated\b', r'\bsystem\b', r'\bnotification\b',
             r'\balert\b', r'\breminder\b', r'\bconfirm\b', r'\bverify\b',
             r'\bunsubscribe\b', r'\bclick here\b', r'\bvisit our\b'
         ]
-        
+
         self.logger.info("Contact Extraction Agent initialized successfully")
-    
+
     @log_function()
-    async def extract_contacts(self, email: EmailMessage) -> List[ContactInfo]:
+    async def extract_contacts(self, email: EmailMessage) -> list[ContactInfo]:
         """
         Main entry point for contact extraction from an email message.
         
@@ -288,43 +288,43 @@ class ContactExtractor:
         """
         if not email or not email.sender:
             raise ValueError("Email and sender information are required")
-        
+
         self.logger.info(f"Extracting contacts from email: {email.sender}")
-        
+
         try:
             # Step 1: Analyze if sender is likely a real person
             person_analysis = await self._analyze_real_person(email)
-            
+
             if person_analysis['confidence'] == ContactConfidence.NONE:
                 self.logger.info(f"Skipping extraction: {person_analysis['reasoning']}")
                 return []
-            
+
             # Step 2: Extract detailed contact information
             contact_info = await self._extract_contact_info(email, person_analysis)
-            
+
             if not contact_info:
                 self.logger.info("No extractable contact information found")
                 return []
-            
+
             # Step 3: Enhance contact with additional analysis
             contact_info = await self._enhance_contact(contact_info, email)
-            
+
             # Step 4: Store successful extraction in memory
             await self._store_extraction(contact_info, email)
-            
+
             self.logger.info(
                 f"Successfully extracted contact: {contact_info.name or contact_info.email} "
                 f"(confidence: {contact_info.confidence.value})"
             )
-            
+
             return [contact_info]
-            
+
         except Exception as e:
             self.logger.error(f"Contact extraction failed for {email.sender}: {e}")
             return []
-    
+
     @log_function()
-    async def _analyze_real_person(self, email: EmailMessage) -> Dict[str, Any]:
+    async def _analyze_real_person(self, email: EmailMessage) -> dict[str, Any]:
         """
         Determine if the email sender is likely a real person vs. automated system.
         
@@ -338,18 +338,18 @@ class ContactExtractor:
             Dictionary containing confidence level and reasoning
         """
         self.logger.debug(f"Analyzing sender authenticity: {email.sender}")
-        
+
         sender_email = email.sender.lower().strip()
-        
+
         # Validate email format
         if '@' not in sender_email:
             return {
                 'confidence': ContactConfidence.NONE,
                 'reasoning': 'Invalid email format'
             }
-        
+
         local_part, domain = sender_email.split('@', 1)
-        
+
         # Check for no-reply patterns
         for pattern in self.no_reply_patterns:
             if re.search(pattern, sender_email, re.IGNORECASE):
@@ -357,21 +357,21 @@ class ContactExtractor:
                     'confidence': ContactConfidence.NONE,
                     'reasoning': f'No-reply address pattern detected: {pattern}'
                 }
-        
+
         # Check for bulk email domains
         if domain in self.bulk_domains:
             return {
                 'confidence': ContactConfidence.NONE,
                 'reasoning': f'Known bulk email domain: {domain}'
             }
-        
+
         # Check for automated system patterns in local part
         automated_indicators = [
             'notification', 'alert', 'system', 'admin', 'support',
             'info', 'sales', 'marketing', 'newsletter', 'updates',
             'service', 'help', 'contact', 'team'
         ]
-        
+
         for indicator in automated_indicators:
             if indicator in local_part:
                 # Still might be a person, check content for personal indicators
@@ -382,29 +382,29 @@ class ContactExtractor:
                         'reasoning': f'Automated system pattern in address: {indicator}'
                     }
                 break
-        
+
         # Check semantic memory for known sender patterns
         sender_knowledge = await self._check_sender_knowledge(email.sender)
-        
+
         if sender_knowledge:
             trust_level = sender_knowledge.get('trust_level', 'unknown')
             sender_type = sender_knowledge.get('sender_type', 'unknown')
-            
+
             if trust_level == 'spam' or trust_level == 'none':
                 return {
                     'confidence': ContactConfidence.NONE,
                     'reasoning': 'Previously identified as spam/untrusted sender'
                 }
-            
+
             if sender_type in ['family', 'colleague', 'friend', 'personal']:
                 return {
                     'confidence': ContactConfidence.HIGH,
                     'reasoning': f'Known trusted contact type: {sender_type}'
                 }
-        
+
         # Calculate personal content score
         personal_score = self._calculate_personal_score(email)
-        
+
         # Determine confidence based on multiple factors
         if personal_score >= 0.7:
             confidence = ContactConfidence.HIGH
@@ -418,15 +418,15 @@ class ContactExtractor:
         else:
             confidence = ContactConfidence.NONE
             reasoning = f'Very low personal content score: {personal_score:.2f}'
-        
+
         self.logger.debug(f"Person analysis result: {confidence.value} - {reasoning}")
-        
+
         return {
             'confidence': confidence,
             'reasoning': reasoning,
             'personal_score': personal_score
         }
-    
+
     @log_function()
     def _calculate_personal_score(self, email: EmailMessage) -> float:
         """
@@ -442,53 +442,53 @@ class ContactExtractor:
             Score between 0.0 (clearly automated) and 1.0 (clearly personal)
         """
         content = f"{email.subject} {email.content}".lower()
-        
+
         # Count personal indicators
         personal_count = sum(
             len(re.findall(pattern, content, re.IGNORECASE))
             for pattern in self.personal_indicators
         )
-        
+
         # Count automated indicators
         automated_count = sum(
             len(re.findall(pattern, content, re.IGNORECASE))
             for pattern in self.automated_indicators
         )
-        
+
         # Check for personal pronouns and conversational language
         personal_pronouns = len(re.findall(r'\b(i|me|my|we|us|our|you|your)\b', content))
         questions = len(re.findall(r'\?', content))
         exclamations = len(re.findall(r'!', content))
-        
+
         # Check for signature-like patterns
         signature_indicators = len(re.findall(
             r'(best regards|sincerely|thanks|cheers|yours|sent from)', content
         ))
-        
+
         # Calculate composite score
         personal_signals = personal_count + personal_pronouns + questions + signature_indicators
         automated_signals = automated_count
-        
+
         # Length normalization
         content_length = max(len(content.split()), 1)
         normalized_personal = min(personal_signals / content_length * 10, 1.0)
         normalized_automated = min(automated_signals / content_length * 10, 1.0)
-        
+
         # Final score calculation
         if normalized_automated > normalized_personal * 2:
             score = max(0.0, 0.3 - normalized_automated * 0.5)
         else:
             score = min(1.0, normalized_personal * 0.7 + 0.3)
-        
+
         self.logger.debug(
             f"Personal score calculation: personal={personal_signals}, "
             f"automated={automated_signals}, score={score:.3f}"
         )
-        
+
         return score
-    
+
     @log_function()
-    async def _check_sender_knowledge(self, sender_email: str) -> Optional[Dict[str, Any]]:
+    async def _check_sender_knowledge(self, sender_email: str) -> dict[str, Any] | None:
         """
         Check semantic memory for existing knowledge about the sender.
         
@@ -507,24 +507,24 @@ class ContactExtractor:
                 query=f"sender:{sender_email}",
                 limit=1
             )
-            
+
             if search_results:
                 memory_item, score = search_results[0]
                 if score > 0.8:  # High confidence match
                     return memory_item.metadata
-            
+
             return None
-            
+
         except Exception as e:
             self.logger.warning(f"Failed to check sender knowledge for {sender_email}: {e}")
             return None
-    
+
     @log_function()
     async def _extract_contact_info(
-        self, 
-        email: EmailMessage, 
-        person_analysis: Dict[str, Any]
-    ) -> Optional[ContactInfo]:
+        self,
+        email: EmailMessage,
+        person_analysis: dict[str, Any]
+    ) -> ContactInfo | None:
         """
         Extract detailed contact information from the email.
         
@@ -539,7 +539,7 @@ class ContactExtractor:
             ContactInfo object with extracted details or None if extraction fails
         """
         self.logger.debug(f"Extracting contact details from {email.sender}")
-        
+
         try:
             # Initialize contact info with email address
             contact = ContactInfo(
@@ -548,22 +548,22 @@ class ContactExtractor:
                 confidence=person_analysis['confidence'],
                 extraction_reasoning=person_analysis['reasoning']
             )
-            
+
             # Extract name from sender field or signature
             contact.name = self._extract_name_from_sender(email.sender)
-            
+
             # Parse signature for additional details
             signature_info = self._extract_from_signature(email.content)
-            
+
             # Merge signature information
             if signature_info.get('name') and not contact.name:
                 contact.name = signature_info['name']
-            
+
             contact.phone = signature_info.get('phone')
             contact.organization = signature_info.get('organization')
             contact.title = signature_info.get('title')
             contact.address = signature_info.get('address')
-            
+
             # Parse name into components
             if contact.name:
                 name_parts = contact.name.strip().split()
@@ -572,10 +572,10 @@ class ContactExtractor:
                     contact.last_name = ' '.join(name_parts[1:])
                 elif len(name_parts) == 1:
                     contact.first_name = name_parts[0]
-            
+
             # Determine contact type
             contact.contact_type = self._determine_contact_type(email, contact)
-            
+
             # Add extraction metadata
             contact.metadata = {
                 'extraction_date': datetime.now(UTC).isoformat(),
@@ -584,20 +584,20 @@ class ContactExtractor:
                 'email_subject': email.subject,
                 'extraction_version': '1.0'
             }
-            
+
             # Validate that we have meaningful information
             if not contact.name and not contact.phone and not contact.organization:
                 self.logger.debug("Insufficient contact information extracted")
                 return None
-            
+
             return contact
-            
+
         except Exception as e:
             self.logger.error(f"Failed to extract contact info: {e}")
             return None
-    
+
     @log_function()
-    def _extract_name_from_sender(self, sender: str) -> Optional[str]:
+    def _extract_name_from_sender(self, sender: str) -> str | None:
         """
         Extract human name from sender field.
         
@@ -611,23 +611,23 @@ class ContactExtractor:
         """
         if not sender:
             return None
-        
+
         # Handle format: "John Doe <john@example.com>"
         if '<' in sender and '>' in sender:
             name_part = sender.split('<')[0].strip()
             # Clean up quotes and whitespace
             name_part = name_part.strip('"\'').strip()
-            if name_part and not '@' in name_part:
+            if name_part and '@' not in name_part:
                 return name_part
-        
+
         # Handle format where email is the display name (skip)
         if '@' in sender:
             return None
-        
+
         return sender.strip() if sender.strip() else None
-    
+
     @log_function()
-    def _extract_from_signature(self, content: str) -> Dict[str, str]:
+    def _extract_from_signature(self, content: str) -> dict[str, str]:
         """
         Extract contact information from email signature.
         
@@ -642,12 +642,12 @@ class ContactExtractor:
         """
         if not content:
             return {}
-        
+
         signature_info = {}
-        
+
         # Look for common signature patterns
         lines = content.split('\n')
-        
+
         # Find potential signature start
         signature_start = -1
         for i, line in enumerate(lines):
@@ -655,21 +655,21 @@ class ContactExtractor:
             if any(marker in line_lower for marker in ['regards', 'sincerely', 'thanks', 'best', '--']):
                 signature_start = i
                 break
-        
+
         if signature_start == -1:
             # No clear signature, look in last few lines
             signature_start = max(0, len(lines) - 5)
-        
+
         signature_lines = lines[signature_start:]
         signature_text = '\n'.join(signature_lines)
-        
+
         # Extract phone numbers
         phone_matches = self.phone_pattern.findall(signature_text)
         if phone_matches:
             # Format first phone number found
             area, exchange, number = phone_matches[0]
             signature_info['phone'] = f"({area}) {exchange}-{number}"
-        
+
         # Extract names (look for capitalized words)
         name_matches = self.name_pattern.findall(signature_text)
         if name_matches:
@@ -678,36 +678,36 @@ class ContactExtractor:
             potential_names = [name for name in name_matches if name not in non_names]
             if potential_names:
                 signature_info['name'] = potential_names[0]
-        
+
         # Look for organization/company names
         org_patterns = [
             r'(?:Company|Corp|Inc|LLC|Ltd)\.?',  # Company suffixes
             r'(?:at\s+)([A-Z][a-zA-Z\s&]+)',     # "at Company Name"
         ]
-        
+
         for pattern in org_patterns:
             matches = re.findall(pattern, signature_text, re.IGNORECASE)
             if matches:
                 signature_info['organization'] = matches[0].strip()
                 break
-        
+
         # Look for job titles
         title_patterns = [
             r'(?:Manager|Director|President|VP|CEO|CTO|CFO|Partner)',
             r'(?:Senior|Principal|Lead)\s+\w+',
             r'(?:Vice President|Executive|Analyst|Associate)',
         ]
-        
+
         for pattern in title_patterns:
             matches = re.findall(pattern, signature_text, re.IGNORECASE)
             if matches:
                 signature_info['title'] = matches[0].strip()
                 break
-        
+
         self.logger.debug(f"Signature extraction found: {list(signature_info.keys())}")
-        
+
         return signature_info
-    
+
     @log_function()
     def _determine_contact_type(self, email: EmailMessage, contact: ContactInfo) -> ContactType:
         """
@@ -723,29 +723,29 @@ class ContactExtractor:
             ContactType enum value
         """
         content_lower = f"{email.subject} {email.content}".lower()
-        
+
         # Family indicators
         family_terms = ['family', 'mom', 'dad', 'sister', 'brother', 'aunt', 'uncle', 'cousin']
         if any(term in content_lower for term in family_terms):
             return ContactType.FAMILY
-        
+
         # indicators
         business_terms = ['meeting', 'project', 'deadline', 'proposal', 'contract', 'business']
         if any(term in content_lower for term in business_terms):
             return ContactType.PROFESSIONAL
-        
+
         # Vendor indicators
         vendor_terms = ['invoice', 'payment', 'service', 'quote', 'delivery', 'order']
         if any(term in content_lower for term in vendor_terms):
             return ContactType.VENDOR
-        
+
         # Check organization for business classification
         if contact.organization:
             return ContactType.PROFESSIONAL
-        
+
         # Default to personal for individual contacts
         return ContactType.PERSONAL
-    
+
     @log_function()
     async def _enhance_contact(self, contact: ContactInfo, email: EmailMessage) -> ContactInfo:
         """
@@ -766,13 +766,13 @@ class ContactExtractor:
                 query=contact.email,
                 limit=1
             )
-            
+
             if existing_contacts:
                 existing_item, score = existing_contacts[0]
                 if score > 0.9:  # Very high match
                     # Merge with existing contact information
                     existing_data = existing_item.metadata
-                    
+
                     # Update with new information if better
                     if not contact.name and existing_data.get('name'):
                         contact.name = existing_data['name']
@@ -780,18 +780,18 @@ class ContactExtractor:
                         contact.phone = existing_data['phone']
                     if not contact.organization and existing_data.get('organization'):
                         contact.organization = existing_data['organization']
-                    
+
                     # Increase confidence if we have historical data
                     if contact.confidence == ContactConfidence.LOW:
                         contact.confidence = ContactConfidence.MEDIUM
                         contact.extraction_reasoning += " (enhanced with historical data)"
-            
+
             return contact
-            
+
         except Exception as e:
             self.logger.warning(f"Failed to enhance contact {contact.email}: {e}")
             return contact
-    
+
     @log_function()
     async def _store_extraction(self, contact: ContactInfo, email: EmailMessage) -> None:
         """
@@ -810,7 +810,7 @@ class ContactExtractor:
                 content=f"Contact: {contact.name or contact.email}",
                 metadata=contact.to_dict()
             )
-            
+
             # Store extraction pattern in procedural memory
             await self.procedural_memory.add(
                 content=f"Extracted contact from {email.sender}",
@@ -823,7 +823,7 @@ class ContactExtractor:
                     'extraction_date': datetime.now(UTC).isoformat()
                 }
             )
-            
+
             # Store sender information in semantic memory
             await self.semantic_memory.add(
                 content=f"Sender: {email.sender} is a real person",
@@ -835,14 +835,14 @@ class ContactExtractor:
                     'last_seen': datetime.now(UTC).isoformat()
                 }
             )
-            
+
             self.logger.debug(f"Stored extraction results for {contact.email}")
-            
+
         except Exception as e:
             self.logger.error(f"Failed to store extraction results: {e}")
-    
+
     @log_function()
-    async def get_extracted_contacts(self, limit: int = 50) -> List[Dict[str, Any]]:
+    async def get_extracted_contacts(self, limit: int = 50) -> list[dict[str, Any]]:
         """
         Retrieve previously extracted contacts from memory.
         
@@ -860,21 +860,21 @@ class ContactExtractor:
                 query="Contact:",
                 limit=limit
             )
-            
+
             contacts = []
             for memory_item, score in search_results:
                 if 'email' in memory_item.metadata:
                     contacts.append(memory_item.metadata)
-            
+
             self.logger.info(f"Retrieved {len(contacts)} extracted contacts")
             return contacts
-            
+
         except Exception as e:
             self.logger.error(f"Failed to retrieve contacts: {e}")
             return []
-    
+
     @log_function()
-    async def search_contacts(self, query: str, limit: int = 10) -> List[Dict[str, Any]]:
+    async def search_contacts(self, query: str, limit: int = 10) -> list[dict[str, Any]]:
         """
         Search extracted contacts by name, email, or organization.
         
@@ -890,29 +890,29 @@ class ContactExtractor:
         """
         if not query or not query.strip():
             return []
-        
+
         try:
             search_results = await self.contact_memory.search(
                 query=query.strip(),
                 limit=limit
             )
-            
+
             contacts = []
             for memory_item, score in search_results:
                 if 'email' in memory_item.metadata:
                     contact_data = memory_item.metadata.copy()
                     contact_data['relevance_score'] = score
                     contacts.append(contact_data)
-            
+
             self.logger.info(f"Contact search '{query}' returned {len(contacts)} results")
             return contacts
-            
+
         except Exception as e:
             self.logger.error(f"Contact search failed for query '{query}': {e}")
             return []
-    
+
     @log_function()
-    async def get_contact_statistics(self) -> Dict[str, Any]:
+    async def get_contact_statistics(self) -> dict[str, Any]:
         """
         Get statistics about extracted contacts.
         
@@ -921,24 +921,24 @@ class ContactExtractor:
         """
         try:
             all_contacts = await self.get_extracted_contacts(limit=1000)
-            
+
             stats = {
                 'total_contacts': len(all_contacts),
                 'by_type': {},
                 'by_confidence': {},
                 'recent_extractions': 0
             }
-            
+
             # Calculate statistics
             for contact in all_contacts:
                 # Count by type
                 contact_type = contact.get('contact_type', 'unknown')
                 stats['by_type'][contact_type] = stats['by_type'].get(contact_type, 0) + 1
-                
+
                 # Count by confidence
                 confidence = contact.get('confidence', 'low')
                 stats['by_confidence'][confidence] = stats['by_confidence'].get(confidence, 0) + 1
-                
+
                 # Count recent extractions (last 7 days)
                 extraction_date = contact.get('metadata', {}).get('extraction_date')
                 if extraction_date:
@@ -949,10 +949,10 @@ class ContactExtractor:
                             stats['recent_extractions'] += 1
                     except:
                         pass
-            
+
             self.logger.info(f"Contact statistics: {stats['total_contacts']} total contacts")
             return stats
-            
+
         except Exception as e:
             self.logger.error(f"Failed to get contact statistics: {e}")
             return {'total_contacts': 0, 'by_type': {}, 'by_confidence': {}, 'recent_extractions': 0}
@@ -966,12 +966,12 @@ async def demo_contact_extractor() -> None:
     Shows how to use the ContactExtractor with sample email data
     and displays the extraction results.
     """
-    print("🤖 Contact Extraction Agent Demo")
-    print("=" * 50)
-    
+    print("🤖 Contact Extraction Agent Demo")  # noqa: T201
+    print("=" * 50)  # noqa: T201
+
     # Initialize extractor
     extractor = ContactExtractor()
-    
+
     # Sample email for demonstration
     sample_email = EmailMessage(
         id="demo_001",
@@ -990,36 +990,36 @@ Phone: (555) 123-4567
 john.smith@acme-corp.com
 """
     )
-    
-    print(f"📧 Analyzing email from: {sample_email.sender}")
-    print(f"📝 Subject: {sample_email.subject}")
-    
+
+    print(f"📧 Analyzing email from: {sample_email.sender}")  # noqa: T201
+    print(f"📝 Subject: {sample_email.subject}")  # noqa: T201
+
     # Extract contacts
     extracted_contacts = await extractor.extract_contacts(sample_email)
-    
+
     if extracted_contacts:
         for contact in extracted_contacts:
-            print(f"\n✅ Contact Extracted:")
-            print(f"   Name: {contact.name}")
-            print(f"   Email: {contact.email}")
-            print(f"   Phone: {contact.phone}")
-            print(f"   Organization: {contact.organization}")
-            print(f"   Title: {contact.title}")
-            print(f"   Type: {contact.contact_type.value}")
-            print(f"   Confidence: {contact.confidence.value}")
-            print(f"   Reasoning: {contact.extraction_reasoning}")
+            print("\n✅ Contact Extracted:")  # noqa: T201
+            print(f"   Name: {contact.name}")  # noqa: T201
+            print(f"   Email: {contact.email}")  # noqa: T201
+            print(f"   Phone: {contact.phone}")  # noqa: T201
+            print(f"   Organization: {contact.organization}")  # noqa: T201
+            print(f"   Title: {contact.title}")  # noqa: T201
+            print(f"   Type: {contact.contact_type.value}")  # noqa: T201
+            print(f"   Confidence: {contact.confidence.value}")  # noqa: T201
+            print(f"   Reasoning: {contact.extraction_reasoning}")  # noqa: T201
     else:
-        print("\n❌ No contacts extracted")
-    
+        print("\n❌ No contacts extracted")  # noqa: T201
+
     # Show statistics
     stats = await extractor.get_contact_statistics()
-    print(f"\n📊 Extraction Statistics:")
-    print(f"   Total contacts: {stats['total_contacts']}")
-    print(f"   Recent extractions: {stats['recent_extractions']}")
-    
-    print("\n✨ Demo completed!")
+    print("\n📊 Extraction Statistics:")  # noqa: T201
+    print(f"   Total contacts: {stats['total_contacts']}")  # noqa: T201
+    print(f"   Recent extractions: {stats['recent_extractions']}")  # noqa: T201
+
+    print("\n✨ Demo completed!")  # noqa: T201
 
 
 if __name__ == "__main__":
     # Run demo if executed directly
-    asyncio.run(demo_contact_extractor()) 
+    asyncio.run(demo_contact_extractor())
