@@ -38,42 +38,43 @@ License -- for Inveniam use only
 Copyright 2025 by Inveniam Capital Partners, LLC and Rick Bunker
 """
 
-import json
-import base64
 import asyncio
-import threading
-import socketserver
+import base64
 import http.server
+import json
+import os
+import socketserver
+
+# Logging system
+import sys
+import threading
 import webbrowser
-from typing import List, Optional, Dict, Any, AsyncGenerator, Union, Callable
-from datetime import datetime, timezone
+from collections.abc import Callable
 from concurrent.futures import ThreadPoolExecutor
+from datetime import UTC, datetime
 from pathlib import Path
-from urllib.parse import urlparse, parse_qs
+from typing import Any
+from urllib.parse import parse_qs, urlparse
 
 import aiohttp
 import msal
 
-# Logging system
-import sys
-import os
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 
-from utils.logging_system import get_logger, log_function, log_debug, log_info
+from utils.logging_system import get_logger, log_function
 
 from .base import (
+    AuthenticationError,
     BaseEmailInterface,
+    ConnectionError,
     Email,
     EmailAddress,
     EmailAttachment,
+    EmailImportance,
+    EmailNotFoundError,
     EmailSearchCriteria,
     EmailSendRequest,
-    EmailImportance,
     EmailSystemError,
-    AuthenticationError,
-    ConnectionError,
-    PermissionError,
-    EmailNotFoundError
 )
 
 # Initialize logger
@@ -91,7 +92,7 @@ class AuthorizationHandler(http.server.SimpleHTTPRequestHandler):
         Provides authentication experience for asset management
         users accessing Microsoft 365 email systems securely.
     """
-    
+
     @log_function()
     def do_GET(self) -> None:
         """
@@ -109,18 +110,18 @@ class AuthorizationHandler(http.server.SimpleHTTPRequestHandler):
                 # Extract authorization code and state from callback
                 parsed = urlparse(self.path)
                 query_params = parse_qs(parsed.query)
-                
+
                 if 'code' in query_params:
                     self.server.auth_code = query_params['code'][0]
                     self.server.state = query_params.get('state', [None])[0]
-                    
+
                     logger.info("OAuth authorization code received successfully")
-                    
+
                     # Send success response
                     self.send_response(200)
                     self.send_header('Content-type', 'text/html')
                     self.end_headers()
-                    
+
                     success_html = """
                     <html>
                     <head>
@@ -151,31 +152,31 @@ class AuthorizationHandler(http.server.SimpleHTTPRequestHandler):
                     # Handle missing authorization code
                     logger.error("OAuth callback missing authorization code")
                     self._send_error_response("No authorization code received")
-                    
+
             elif self.path.startswith('/?error='):
                 # Handle OAuth errors
                 parsed = urlparse(self.path)
                 query_params = parse_qs(parsed.query)
                 error_code = query_params.get('error', ['unknown'])[0]
                 error_description = query_params.get('error_description', ['Unknown error'])[0]
-                
+
                 logger.error(f"OAuth error: {error_code} - {error_description}")
                 self._send_error_response(f"Authentication failed: {error_description}")
-                
+
             else:
                 # Default waiting response
                 self._send_waiting_response()
-                
+
         except Exception as e:
             logger.error(f"OAuth callback handler error: {e}")
             self._send_error_response("Authentication processing error")
-    
+
     def _send_error_response(self, error_message: str) -> None:
         """Send error response to user."""
         self.send_response(400)
         self.send_header('Content-type', 'text/html')
         self.end_headers()
-        
+
         error_html = f"""
         <html>
         <head>
@@ -201,13 +202,13 @@ class AuthorizationHandler(http.server.SimpleHTTPRequestHandler):
         </html>
         """
         self.wfile.write(error_html.encode())
-    
+
     def _send_waiting_response(self) -> None:
         """Send waiting response to user."""
         self.send_response(200)
         self.send_header('Content-type', 'text/html')
         self.end_headers()
-        
+
         waiting_html = """
         <html>
         <head>
@@ -235,7 +236,7 @@ class AuthorizationHandler(http.server.SimpleHTTPRequestHandler):
         </html>
         """
         self.wfile.write(waiting_html.encode())
-    
+
     def log_message(self, format: str, *args: Any) -> None:
         """Suppress default HTTP server logging to avoid noise."""
         pass
@@ -268,23 +269,23 @@ class MicrosoftGraphInterface(BaseEmailInterface):
         - Complete error mapping and handling
         - logging and monitoring
     """
-    
+
     # Microsoft Graph API configuration
     SCOPES = [
         'https://graph.microsoft.com/Mail.ReadWrite',
         'https://graph.microsoft.com/Mail.Send',
         'https://graph.microsoft.com/User.Read'
     ]
-    
+
     GRAPH_ENDPOINT = 'https://graph.microsoft.com/v1.0'
     AUTH_ENDPOINT = 'https://login.microsoftonline.com'
-    
+
     # API limits and timeouts
     DEFAULT_PAGE_SIZE = 25
     MAX_PAGE_SIZE = 100
     REQUEST_TIMEOUT = 30
     AUTH_TIMEOUT = 300  # 5 minutes for authentication
-    
+
     def __init__(self, credentials_path: str = "config/msgraph_credentials.json") -> None:
         """
         Initialize Microsoft Graph interface with complete configuration.
@@ -306,29 +307,29 @@ class MicrosoftGraphInterface(BaseEmailInterface):
             appropriate permissions for email access in the target tenant.
         """
         super().__init__()
-        
+
         self.logger = get_logger(f"{__name__}.MicrosoftGraphInterface")
         self.logger.info("Initializing Microsoft Graph email interface")
-        
+
         # Core components
-        self.access_token: Optional[str] = None
-        self.app: Optional[msal.PublicClientApplication] = None
-        self.session: Optional[aiohttp.ClientSession] = None
+        self.access_token: str | None = None
+        self.app: msal.PublicClientApplication | None = None
+        self.session: aiohttp.ClientSession | None = None
         self.executor = ThreadPoolExecutor(max_workers=4, thread_name_prefix="msgraph")
-        
+
         # Load and validate credentials
         self.credentials_path = Path(credentials_path)
         self.credentials = self._load_credentials()
-        
+
         # Performance tracking
         self.request_count = 0
         self.error_count = 0
         self.total_request_time = 0.0
-        
+
         self.logger.info(f"Microsoft Graph interface initialized for {self.credentials.get('application_name', 'Unknown App')}")
-    
+
     @log_function()
-    def _load_credentials(self) -> Dict[str, Any]:
+    def _load_credentials(self) -> dict[str, Any]:
         """
         Load and validate Microsoft Graph credentials from file.
         
@@ -344,33 +345,33 @@ class MicrosoftGraphInterface(BaseEmailInterface):
         """
         if not self.credentials_path.exists():
             raise FileNotFoundError(f"Microsoft Graph credentials file not found: {self.credentials_path}")
-        
+
         try:
-            with open(self.credentials_path, 'r') as f:
+            with open(self.credentials_path) as f:
                 credentials = json.load(f)
-            
+
             # Validate required fields
             required_fields = ['client_id', 'tenant_id']
             missing_fields = [field for field in required_fields if field not in credentials]
-            
+
             if missing_fields:
                 raise ValueError(f"Missing required credential fields: {missing_fields}")
-            
+
             # Validate field formats
             if not credentials['client_id'] or len(credentials['client_id']) < 10:
                 raise ValueError("Invalid client_id in credentials")
-            
+
             if not credentials['tenant_id'] or len(credentials['tenant_id']) < 10:
                 raise ValueError("Invalid tenant_id in credentials")
-            
+
             self.logger.info(f"Loaded Microsoft Graph credentials for tenant: {credentials['tenant_id'][:8]}...")
             return credentials
-            
+
         except json.JSONDecodeError as e:
             raise ValueError(f"Invalid JSON in credentials file: {e}")
         except Exception as e:
             raise ValueError(f"Error loading credentials: {e}")
-    
+
     @log_function()
     async def connect(self) -> bool:
         """
@@ -397,67 +398,67 @@ class MicrosoftGraphInterface(BaseEmailInterface):
         """
         try:
             self.logger.info("Starting Microsoft Graph authentication flow")
-            
+
             # Extract credentials
             client_id = self.credentials['client_id']
             tenant_id = self.credentials['tenant_id']
             redirect_uri = "http://localhost:8080"
-            
+
             authority = f"{self.AUTH_ENDPOINT}/{tenant_id}"
-            
+
             # Create MSAL public client application
             self.app = msal.PublicClientApplication(
                 client_id=client_id,
                 authority=authority
             )
-            
+
             # Check for cached tokens first
             accounts = self.app.get_accounts()
             token_result = None
-            
+
             if accounts:
                 self.logger.info(f"Found {len(accounts)} cached Microsoft account(s)")
                 # Attempt silent token acquisition
                 token_result = self.app.acquire_token_silent(self.SCOPES, account=accounts[0])
-                
+
                 if token_result:
                     self.logger.info("Successfully acquired token silently (cached/refreshed)")
                 else:
                     self.logger.info("Silent token acquisition failed - interactive authentication required")
-            
+
             # If no cached token or silent acquisition failed, do interactive flow
             if not token_result:
                 token_result = await self._perform_interactive_authentication(redirect_uri)
-            
+
             # Validate token result
             if not token_result or 'access_token' not in token_result:
                 error_msg = token_result.get('error_description', 'Unknown authentication error') if token_result else 'No token received'
                 raise AuthenticationError(f"Microsoft Graph authentication failed: {error_msg}")
-            
+
             # Store access token
             self.access_token = token_result['access_token']
             self.logger.info("Microsoft Graph access token acquired successfully")
-            
+
             # Initialize HTTP session
             await self._initialize_http_session()
-            
+
             # Validate connection by getting user profile
             profile = await self.get_profile()
             self.user_email = profile.get('email')
             self.display_name = profile.get('name')
             self.is_connected = True
-            
+
             self.logger.info(f"Microsoft Graph connection established for {self.user_email}")
             return True
-            
+
         except AuthenticationError:
             raise
         except Exception as e:
             self.logger.error(f"Microsoft Graph connection failed: {e}")
             raise ConnectionError(f"Failed to connect to Microsoft Graph: {e}")
-    
+
     @log_function()
-    async def _perform_interactive_authentication(self, redirect_uri: str) -> Optional[Dict[str, Any]]:
+    async def _perform_interactive_authentication(self, redirect_uri: str) -> dict[str, Any] | None:
         """
         Perform interactive OAuth 2.0 authentication with web browser.
         
@@ -476,12 +477,12 @@ class MicrosoftGraphInterface(BaseEmailInterface):
             RuntimeError: If local server cannot be started
         """
         self.logger.info("Starting interactive Microsoft Graph authentication")
-        
+
         # Start local callback server
         port = 8080
         max_port_attempts = 10
         server = None
-        
+
         for attempt in range(max_port_attempts):
             try:
                 server = socketserver.TCPServer(("", port + attempt), AuthorizationHandler)
@@ -493,76 +494,76 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                 if attempt == max_port_attempts - 1:
                     raise RuntimeError("Cannot start local server for OAuth callback")
                 continue
-        
+
         # Start server in background thread
         server_thread = threading.Thread(target=server.serve_forever, daemon=True)
         server_thread.start()
-        
+
         try:
             # Update redirect URI with actual port
             actual_redirect_uri = f"http://localhost:{port}"
-            
+
             # Initialize authorization code flow
             flow = self.app.initiate_auth_code_flow(
                 scopes=self.SCOPES,
                 redirect_uri=actual_redirect_uri
             )
-            
+
             if "auth_uri" not in flow:
                 raise AuthenticationError("Failed to create Microsoft Graph authorization flow")
-            
+
             auth_url = flow["auth_uri"]
-            
+
             # Open browser for authentication
             self.logger.info("Opening browser for Microsoft 365 authentication")
-            print(f"🌐 Opening browser for Microsoft 365 authentication...")
-            print(f"   If browser doesn't open automatically, visit: {auth_url}")
-            
+            logger.info("🌐 Opening browser for Microsoft 365 authentication...")
+            logger.info("   If browser doesn't open automatically, visit: %s", auth_url)
+
             webbrowser.open(auth_url)
-            
+
             # Wait for authorization code
             self.logger.info(f"Waiting for OAuth callback on port {port}")
-            print(f"   Waiting for authentication completion...")
-            
+            logger.info("   Waiting for authentication completion...")
+
             # Poll for authorization code with timeout
             timeout_seconds = self.AUTH_TIMEOUT
             poll_interval = 1
             elapsed = 0
-            
+
             while elapsed < timeout_seconds:
                 if server.auth_code:
                     self.logger.info("Authorization code received from OAuth callback")
                     break
-                
+
                 await asyncio.sleep(poll_interval)
                 elapsed += poll_interval
-                
+
                 if elapsed % 30 == 0:  # Progress update every 30 seconds
                     self.logger.info(f"Still waiting for authentication... ({elapsed}s elapsed)")
-            
+
             if not server.auth_code:
                 raise AuthenticationError(f"Authentication timeout after {timeout_seconds} seconds")
-            
+
             # Complete the flow with authorization code and state
             auth_response = {'code': server.auth_code}
             if server.state:
                 auth_response['state'] = server.state
-            
+
             token_result = self.app.acquire_token_by_auth_code_flow(
                 auth_code_flow=flow,
                 auth_response=auth_response
             )
-            
+
             self.logger.info("Interactive authentication completed successfully")
-            print("✅ Microsoft 365 authentication successful!")
-            
+            logger.info("✅ Microsoft 365 authentication successful!")
+
             return token_result
-            
+
         finally:
             # Clean up server
             server.shutdown()
             server.server_close()
-    
+
     @log_function()
     async def _initialize_http_session(self) -> None:
         """
@@ -573,24 +574,24 @@ class MicrosoftGraphInterface(BaseEmailInterface):
         """
         if self.session:
             await self.session.close()
-        
+
         # Configure session with authentication headers
         headers = {
             'Authorization': f'Bearer {self.access_token}',
             'Content-Type': 'application/json',
             'User-Agent': 'EmailAgent-MicrosoftGraph/1.0'
         }
-        
+
         timeout = aiohttp.ClientTimeout(total=self.REQUEST_TIMEOUT)
-        
+
         self.session = aiohttp.ClientSession(
             headers=headers,
             timeout=timeout,
             raise_for_status=False  # Handle errors manually
         )
-        
+
         self.logger.info("HTTP session initialized for Microsoft Graph API")
-    
+
     @log_function()
     async def disconnect(self) -> None:
         """
@@ -600,28 +601,28 @@ class MicrosoftGraphInterface(BaseEmailInterface):
         and resets connection state with complete cleanup.
         """
         self.logger.info("Disconnecting from Microsoft Graph")
-        
+
         # Close HTTP session
         if self.session:
             await self.session.close()
             self.session = None
-        
+
         # Clear authentication
         self.access_token = None
         self.app = None
-        
+
         # Reset connection state
         self.is_connected = False
         self.user_email = None
         self.display_name = None
-        
+
         # Shutdown thread pool
         self.executor.shutdown(wait=True)
-        
+
         self.logger.info("Microsoft Graph disconnection complete")
-    
+
     @log_function()
-    async def get_profile(self) -> Dict[str, Any]:
+    async def get_profile(self) -> dict[str, Any]:
         """
         Retrieve user profile information from Microsoft Graph.
         
@@ -637,15 +638,15 @@ class MicrosoftGraphInterface(BaseEmailInterface):
         """
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             # Get user profile from Microsoft Graph
             url = f"{self.GRAPH_ENDPOINT}/me"
-            
+
             async with self.session.get(url) as response:
                 if response.status == 200:
                     profile_data = await response.json()
-                    
+
                     # Extract relevant profile information
                     profile = {
                         'email': profile_data.get('mail') or profile_data.get('userPrincipalName'),
@@ -655,30 +656,30 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                         'department': profile_data.get('department'),
                         'office_location': profile_data.get('officeLocation')
                     }
-                    
+
                     self.logger.info(f"Retrieved profile for {profile['email']}")
                     return profile
-                    
+
                 else:
                     error_data = await response.json() if response.content_type == 'application/json' else {}
                     error_msg = error_data.get('error', {}).get('message', f'HTTP {response.status}')
                     raise EmailSystemError(f"Failed to get Microsoft Graph profile: {error_msg}")
-                    
+
         except Exception as e:
             if isinstance(e, EmailSystemError):
                 raise
             self.logger.error(f"Microsoft Graph profile retrieval failed: {e}")
             raise EmailSystemError(f"Profile retrieval error: {e}")
-    
-    async def list_emails(self, criteria: EmailSearchCriteria) -> List[Email]:
+
+    async def list_emails(self, criteria: EmailSearchCriteria) -> list[Email]:
         """List emails matching criteria."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             # Build Microsoft Graph filter and search query
             filters = []
-            
+
             if criteria.sender:
                 filters.append(f"from/emailAddress/address eq '{criteria.sender}'")
             if criteria.subject:
@@ -695,34 +696,32 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                 # Use proper ISO format for Microsoft Graph compatibility
                 if criteria.date_after.tzinfo is None:
                     # Add UTC timezone if naive datetime
-                    from datetime import timezone
-                    criteria.date_after = criteria.date_after.replace(tzinfo=timezone.utc)
+                    criteria.date_after = criteria.date_after.replace(tzinfo=UTC)
                 iso_date = criteria.date_after.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
                 filters.append(f"receivedDateTime ge {iso_date}")
             if criteria.date_before:
                 # Use proper ISO format for Microsoft Graph compatibility
                 if criteria.date_before.tzinfo is None:
                     # Add UTC timezone if naive datetime
-                    from datetime import timezone
-                    criteria.date_before = criteria.date_before.replace(tzinfo=timezone.utc)
+                    criteria.date_before = criteria.date_before.replace(tzinfo=UTC)
                 iso_date = criteria.date_before.strftime('%Y-%m-%dT%H:%M:%S.%fZ')
                 filters.append(f"receivedDateTime le {iso_date}")
-            
+
             # Build URL
             url = f"{self.GRAPH_ENDPOINT}/me/messages"
             params = {
                 '$top': str(criteria.max_results),
                 '$orderby': 'receivedDateTime desc'
             }
-            
+
             # If searching for attachments, expand to include attachment data
             if criteria.has_attachments:
                 params['$expand'] = 'attachments'
-            
+
             if filters:
                 # Microsoft Graph requires: properties in $orderby must also appear in $filter
                 # AND properties in $orderby must appear FIRST in $filter
-                
+
                 # Check if we already have receivedDateTime filters from date criteria
                 has_date_filter = any('receivedDateTime' in f for f in filters)
                 if not has_date_filter:
@@ -734,28 +733,28 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                     date_filters = [f for f in filters if 'receivedDateTime' in f]
                     other_filters = [f for f in filters if 'receivedDateTime' not in f]
                     all_filters = date_filters + other_filters
-                
+
                 params['$filter'] = ' and '.join(all_filters)
-            
+
             if criteria.query:
                 params['$search'] = f'"{criteria.query}"'
-            
+
             # Handle labels (folders)
             if criteria.labels:
                 # For Microsoft Graph, we need to search in specific folders
                 # This is a simplified approach - real implementation might need folder hierarchy
                 folder_name = criteria.labels[0]  # Use first label as folder
                 url = f"{self.GRAPH_ENDPOINT}/me/mailFolders('{folder_name}')/messages"
-            
+
             async with self.session.get(url, params=params) as response:
                 if response.status == 401:
                     raise AuthenticationError("Microsoft Graph token expired or invalid")
                 elif response.status != 200:
                     raise EmailSystemError(f"Failed to list messages: HTTP {response.status}")
-                
+
                 data = await response.json()
                 messages = data.get('value', [])
-                
+
                 # Convert to Email objects
                 emails = []
                 include_attachments = criteria.has_attachments  # Include attachments if we searched for them
@@ -764,26 +763,26 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                         email_obj = self._parse_graph_message(msg, include_attachments)
                         emails.append(email_obj)
                     except Exception as e:
-                        print(f"[WARN] Failed to parse message {msg.get('id')}: {e}")
+                        logger.warning("Failed to parse message %s: %s", msg.get('id'), e)
                         continue
-                
+
                 return emails
-                
+
         except aiohttp.ClientError as e:
             raise EmailSystemError(f"Failed to list Microsoft Graph messages: {e}")
-    
+
     async def get_email(self, email_id: str, include_attachments: bool = False) -> Email:
         """Get a specific email by ID."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             # Get message with attachments if requested
             expand = '$expand=attachments' if include_attachments else ''
             url = f"{self.GRAPH_ENDPOINT}/me/messages/{email_id}"
             if expand:
                 url += f"?{expand}"
-            
+
             async with self.session.get(url) as response:
                 if response.status == 401:
                     raise AuthenticationError("Microsoft Graph token expired or invalid")
@@ -791,18 +790,18 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                     raise EmailNotFoundError(f"Email {email_id} not found")
                 elif response.status != 200:
                     raise EmailSystemError(f"Failed to get message: HTTP {response.status}")
-                
+
                 data = await response.json()
                 return self._parse_graph_message(data, include_attachments)
-                
+
         except aiohttp.ClientError as e:
             raise EmailSystemError(f"Failed to get Microsoft Graph message: {e}")
-    
+
     async def send_email(self, request: EmailSendRequest) -> str:
         """Send an email via Microsoft Graph."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             # Build message
             message = {
@@ -820,7 +819,7 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                     'content': request.body_html or request.body_text or ''
                 }
             }
-            
+
             if request.cc:
                 message['ccRecipients'] = [
                     {
@@ -830,7 +829,7 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                         }
                     } for addr in request.cc
                 ]
-            
+
             if request.bcc:
                 message['bccRecipients'] = [
                     {
@@ -840,7 +839,7 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                         }
                     } for addr in request.bcc
                 ]
-            
+
             # Set importance
             if request.importance == EmailImportance.HIGH:
                 message['importance'] = 'high'
@@ -848,7 +847,7 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                 message['importance'] = 'low'
             else:
                 message['importance'] = 'normal'
-            
+
             # Handle reply threading
             if request.reply_to_message_id:
                 # For replies, we should use the reply endpoint instead
@@ -857,7 +856,7 @@ class MicrosoftGraphInterface(BaseEmailInterface):
             else:
                 url = f"{self.GRAPH_ENDPOINT}/me/sendMail"
                 payload = {'message': message}
-            
+
             # Add attachments
             if request.attachments:
                 message['attachments'] = []
@@ -870,37 +869,37 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                             'contentType': attachment.content_type,
                             'contentBytes': content_bytes
                         })
-            
+
             async with self.session.post(url, json=payload) as response:
                 if response.status == 401:
                     raise AuthenticationError("Microsoft Graph token expired or invalid")
                 elif response.status not in [200, 202]:
                     error_text = await response.text()
                     raise EmailSystemError(f"Failed to send message: HTTP {response.status} - {error_text}")
-                
+
                 # Microsoft Graph doesn't return message ID on send
                 # Return a placeholder ID
                 return "sent"
-                
+
         except aiohttp.ClientError as e:
             raise EmailSystemError(f"Failed to send Microsoft Graph message: {e}")
-    
+
     async def mark_as_read(self, email_id: str) -> bool:
         """Mark email as read."""
         return await self._update_message(email_id, {'isRead': True})
-    
+
     async def mark_as_unread(self, email_id: str) -> bool:
         """Mark email as unread."""
         return await self._update_message(email_id, {'isRead': False})
-    
+
     async def delete_email(self, email_id: str) -> bool:
         """Delete an email."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             url = f"{self.GRAPH_ENDPOINT}/me/messages/{email_id}"
-            
+
             async with self.session.delete(url) as response:
                 if response.status == 401:
                     raise AuthenticationError("Microsoft Graph token expired or invalid")
@@ -908,50 +907,50 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                     raise EmailNotFoundError(f"Email {email_id} not found")
                 elif response.status != 204:
                     raise EmailSystemError(f"Failed to delete message: HTTP {response.status}")
-                
+
                 return True
-                
+
         except aiohttp.ClientError as e:
             raise EmailSystemError(f"Failed to delete Microsoft Graph message: {e}")
-    
-    async def get_labels(self) -> List[str]:
+
+    async def get_labels(self) -> list[str]:
         """Get available folders (labels)."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             url = f"{self.GRAPH_ENDPOINT}/me/mailFolders"
-            
+
             async with self.session.get(url) as response:
                 if response.status == 401:
                     raise AuthenticationError("Microsoft Graph token expired or invalid")
                 elif response.status != 200:
                     raise EmailSystemError(f"Failed to get folders: HTTP {response.status}")
-                
+
                 data = await response.json()
                 folders = data.get('value', [])
                 return [folder['displayName'] for folder in folders]
-                
+
         except aiohttp.ClientError as e:
             raise EmailSystemError(f"Failed to get Microsoft Graph folders: {e}")
-    
+
     async def add_label(self, email_id: str, label: str) -> bool:
         """Move email to folder (Microsoft Graph doesn't have labels like Gmail)."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             # First, find the folder ID for the label
             folders = await self._get_folders_dict()
             folder_id = folders.get(label)
-            
+
             if not folder_id:
                 raise EmailSystemError(f"Folder '{label}' not found")
-            
+
             # Move message to folder
             url = f"{self.GRAPH_ENDPOINT}/me/messages/{email_id}/move"
             payload = {'destinationId': folder_id}
-            
+
             async with self.session.post(url, json=payload) as response:
                 if response.status == 401:
                     raise AuthenticationError("Microsoft Graph token expired or invalid")
@@ -959,25 +958,25 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                     raise EmailNotFoundError(f"Email {email_id} not found")
                 elif response.status not in [200, 201]:
                     raise EmailSystemError(f"Failed to move message: HTTP {response.status}")
-                
+
                 return True
-                
+
         except aiohttp.ClientError as e:
             raise EmailSystemError(f"Failed to add label to Microsoft Graph message: {e}")
-    
+
     async def remove_label(self, email_id: str, label: str) -> bool:
         """Move email from folder back to inbox."""
         # For simplicity, we move back to inbox when removing a "label"
         return await self.add_label(email_id, "Inbox")
-    
+
     async def download_attachment(self, email_id: str, attachment_id: str) -> bytes:
         """Download attachment content from Microsoft Graph."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             url = f"{self.GRAPH_ENDPOINT}/me/messages/{email_id}/attachments/{attachment_id}/$value"
-            
+
             async with self.session.get(url) as response:
                 if response.status == 401:
                     raise AuthenticationError("Microsoft Graph token expired or invalid")
@@ -985,27 +984,27 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                     raise EmailNotFoundError(f"Attachment {attachment_id} not found in email {email_id}")
                 elif response.status != 200:
                     raise EmailSystemError(f"Failed to download attachment: HTTP {response.status}")
-                
+
                 content = await response.read()
                 return content
-                
+
         except aiohttp.ClientError as e:
             raise EmailSystemError(f"Failed to download Microsoft Graph attachment: {e}")
-    
+
     # Helper methods
     async def _run_in_executor(self, func: Callable, *args: Any, **kwargs: Any) -> Any:
         """Run blocking function in thread executor."""
         loop = asyncio.get_event_loop()
         return await loop.run_in_executor(self.executor, lambda: func(*args, **kwargs))
-    
-    async def _update_message(self, email_id: str, update_data: Dict[str, Any]) -> bool:
+
+    async def _update_message(self, email_id: str, update_data: dict[str, Any]) -> bool:
         """Update message properties."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             url = f"{self.GRAPH_ENDPOINT}/me/messages/{email_id}"
-            
+
             async with self.session.patch(url, json=update_data) as response:
                 if response.status == 401:
                     raise AuthenticationError("Microsoft Graph token expired or invalid")
@@ -1013,34 +1012,34 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                     raise EmailNotFoundError(f"Email {email_id} not found")
                 elif response.status != 200:
                     raise EmailSystemError(f"Failed to update message: HTTP {response.status}")
-                
+
                 return True
-                
+
         except aiohttp.ClientError as e:
             raise EmailSystemError(f"Failed to update Microsoft Graph message: {e}")
-    
-    async def _get_folders_dict(self) -> Dict[str, str]:
+
+    async def _get_folders_dict(self) -> dict[str, str]:
         """Get folder name to ID mapping."""
         if not self.session:
             raise ConnectionError("Not connected to Microsoft Graph")
-        
+
         try:
             url = f"{self.GRAPH_ENDPOINT}/me/mailFolders"
-            
+
             async with self.session.get(url) as response:
                 if response.status != 200:
                     return {}
-                
+
                 data = await response.json()
                 folders = data.get('value', [])
                 return {folder['displayName']: folder['id'] for folder in folders}
-                
+
         except aiohttp.ClientError:
             return {}
-    
-    def _parse_graph_message(self, message: Dict[str, Any], include_attachments: bool = False) -> Email:
+
+    def _parse_graph_message(self, message: dict[str, Any], include_attachments: bool = False) -> Email:
         """Parse Microsoft Graph message into Email object."""
-        
+
         # Extract basic fields
         subject = message.get('subject', '')
         sender_data = message.get('from', {}).get('emailAddress', {})
@@ -1048,7 +1047,7 @@ class MicrosoftGraphInterface(BaseEmailInterface):
             address=sender_data.get('address', ''),
             name=sender_data.get('name')
         )
-        
+
         # Parse recipients
         recipients = []
         for recipient in message.get('toRecipients', []):
@@ -1057,7 +1056,7 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                 address=addr_data.get('address', ''),
                 name=addr_data.get('name')
             ))
-        
+
         cc = []
         for recipient in message.get('ccRecipients', []):
             addr_data = recipient.get('emailAddress', {})
@@ -1065,26 +1064,26 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                 address=addr_data.get('address', ''),
                 name=addr_data.get('name')
             ))
-        
+
         # Parse dates
         sent_date = None
         received_date = None
-        
+
         if message.get('sentDateTime'):
             sent_date = datetime.fromisoformat(message['sentDateTime'].replace('Z', '+00:00'))
         if message.get('receivedDateTime'):
             received_date = datetime.fromisoformat(message['receivedDateTime'].replace('Z', '+00:00'))
-        
+
         # Parse body
         body_data = message.get('body', {})
         body_text = None
         body_html = None
-        
+
         if body_data.get('contentType') == 'text':
             body_text = body_data.get('content')
         elif body_data.get('contentType') == 'html':
             body_html = body_data.get('content')
-        
+
         # Parse attachments
         attachments = []
         if include_attachments and message.get('attachments'):
@@ -1095,16 +1094,16 @@ class MicrosoftGraphInterface(BaseEmailInterface):
                     size=att.get('size', 0),
                     attachment_id=att.get('id')
                 )
-                
+
                 # If content is included
                 if att.get('contentBytes'):
                     try:
                         attachment.content = base64.b64decode(att['contentBytes'])
                     except:
                         pass
-                
+
                 attachments.append(attachment)
-        
+
         # Parse importance
         importance = EmailImportance.NORMAL
         graph_importance = message.get('importance', 'normal').lower()
@@ -1112,11 +1111,11 @@ class MicrosoftGraphInterface(BaseEmailInterface):
             importance = EmailImportance.HIGH
         elif graph_importance == 'low':
             importance = EmailImportance.LOW
-        
+
         # Parse flags
         is_read = message.get('isRead', False)
         is_flagged = message.get('flag', {}).get('flagStatus') == 'flagged'
-        
+
         # Build headers dict from available data
         headers = {
             'subject': subject,
@@ -1125,10 +1124,10 @@ class MicrosoftGraphInterface(BaseEmailInterface):
             'message-id': message.get('internetMessageId'),
             'date': message.get('sentDateTime', ''),
         }
-        
+
         if message.get('conversationId'):
             headers['conversation-id'] = message['conversationId']
-        
+
         return Email(
             id=message['id'],
             thread_id=message.get('conversationId'),
@@ -1149,4 +1148,4 @@ class MicrosoftGraphInterface(BaseEmailInterface):
             message_id=message.get('internetMessageId'),
             in_reply_to=None,  # Would need to parse this from headers
             raw_data=message
-        ) 
+        )
