@@ -39,20 +39,17 @@ License -- for Inveniam use only
 Copyright 2025 by Inveniam Capital Partners, LLC and Rick Bunker
 """
 
-import time
-import uuid
-from typing import Any, Dict, List, Optional, Union
-from datetime import datetime, UTC
-from enum import Enum
-
-from qdrant_client.http import models
+import os
 
 # Logging system
 import sys
-import os
+from datetime import UTC, datetime
+from enum import Enum
+from typing import Any
+
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 
-from utils.logging_system import get_logger, log_function, log_debug, log_info
+from utils.logging_system import get_logger, log_function
 
 from .base import BaseMemory, MemoryItem
 
@@ -134,10 +131,10 @@ class SemanticMemory(BaseMemory):
         - Knowledge lifecycle management and updates
         - audit trail and knowledge tracking
     """
-    
+
     def __init__(
         self,
-        max_items: Optional[int] = 1000,
+        max_items: int | None = 1000,
         qdrant_url: str = "http://localhost:6333",
         embedding_model: str = "all-MiniLM-L6-v2"
     ):
@@ -156,12 +153,12 @@ class SemanticMemory(BaseMemory):
         )
         self.collection_name = "semantic"
         logger.info(f"Initialized SemanticMemory with max_items={max_items}")
-    
+
     @log_function()
     async def add(
         self,
         content: str,
-        metadata: Optional[Dict[str, Any]] = None,
+        metadata: dict[str, Any] | None = None,
         knowledge_type: KnowledgeType = KnowledgeType.UNKNOWN,
         confidence: KnowledgeConfidence = KnowledgeConfidence.MEDIUM
     ) -> str:
@@ -198,10 +195,10 @@ class SemanticMemory(BaseMemory):
         """
         if not content or not isinstance(content, str):
             raise ValueError("Knowledge content must be a non-empty string")
-        
+
         if metadata is None:
             metadata = {}
-        
+
         # Set complete knowledge metadata
         metadata["type"] = "knowledge"
         metadata["knowledge_type"] = knowledge_type.value
@@ -209,10 +206,10 @@ class SemanticMemory(BaseMemory):
         metadata["created_at"] = datetime.now(UTC).isoformat()
         metadata["source"] = "semantic_memory"
         metadata["version"] = "1.0.0"
-        
+
         logger.info(f"Adding semantic knowledge: type={knowledge_type.value}, confidence={confidence.value}")
         logger.debug(f"Knowledge content length: {len(content)}")
-        
+
         try:
             result = await super().add(content, metadata)
             logger.info(f"Successfully added semantic knowledge: {result}")
@@ -220,15 +217,15 @@ class SemanticMemory(BaseMemory):
         except Exception as e:
             logger.error(f"Failed to add semantic knowledge: {e}")
             raise
-    
+
     @log_function()
     async def add_sender_knowledge(
         self,
         content: str,
         sender_email: str,
-        sender_domain: Optional[str] = None,
-        organization: Optional[str] = None,
-        response_pattern: Optional[str] = None,
+        sender_domain: str | None = None,
+        organization: str | None = None,
+        response_pattern: str | None = None,
         confidence: KnowledgeConfidence = KnowledgeConfidence.MEDIUM,
         **kwargs
     ) -> str:
@@ -268,17 +265,17 @@ class SemanticMemory(BaseMemory):
             "response_pattern": response_pattern,
             **kwargs
         }
-        
+
         logger.info(f"Adding sender knowledge for: {sender_email}")
         return await self.add(content, metadata, KnowledgeType.SENDER, confidence)
-    
+
     @log_function()
     async def add_email_type_knowledge(
         self,
         content: str,
         email_type: str,
-        classification_criteria: Optional[str] = None,
-        typical_patterns: Optional[List[str]] = None,
+        classification_criteria: str | None = None,
+        typical_patterns: list[str] | None = None,
         confidence: KnowledgeConfidence = KnowledgeConfidence.MEDIUM,
         **kwargs
     ) -> str:
@@ -306,17 +303,17 @@ class SemanticMemory(BaseMemory):
             "typical_patterns": typical_patterns or [],
             **kwargs
         }
-        
+
         logger.info(f"Adding email type knowledge: {email_type}")
         return await self.add(content, metadata, KnowledgeType.EMAIL_TYPE, confidence)
-    
+
     @log_function()
     async def add_domain_knowledge(
         self,
         content: str,
         domain: str,
         expertise_area: str,
-        industry_context: Optional[str] = None,
+        industry_context: str | None = None,
         confidence: KnowledgeConfidence = KnowledgeConfidence.MEDIUM,
         **kwargs
     ) -> str:
@@ -344,19 +341,19 @@ class SemanticMemory(BaseMemory):
             "industry_context": industry_context,
             **kwargs
         }
-        
+
         logger.info(f"Adding domain knowledge: {domain}/{expertise_area}")
         return await self.add(content, metadata, KnowledgeType.DOMAIN, confidence)
-    
+
     @log_function()
     async def search(
         self,
         query: str,
         limit: int = 5,
-        filter: Optional[Dict[str, Any]] = None,
-        knowledge_type: Optional[KnowledgeType] = None,
-        min_confidence: Optional[KnowledgeConfidence] = None
-    ) -> List[MemoryItem]:
+        filter: dict[str, Any] | None = None,
+        knowledge_type: KnowledgeType | None = None,
+        min_confidence: KnowledgeConfidence | None = None
+    ) -> list[MemoryItem]:
         """
         Search semantic knowledge with complete filtering.
         
@@ -383,7 +380,7 @@ class SemanticMemory(BaseMemory):
             ... )
         """
         logger.info(f"Searching semantic knowledge: query='{query}', limit={limit}")
-        
+
         try:
             # Build complete filter conditions
             filter_conditions = {
@@ -391,50 +388,50 @@ class SemanticMemory(BaseMemory):
                     {"key": "metadata.type", "match": {"value": "knowledge"}}
                 ]
             }
-            
+
             # Knowledge type filter
             if knowledge_type:
                 filter_conditions["must"].append({
                     "key": "metadata.knowledge_type",
                     "match": {"value": knowledge_type.value}
                 })
-            
+
             # Minimum confidence filter
             if min_confidence:
                 confidence_values = [conf.value for conf in KnowledgeConfidence]
                 min_index = confidence_values.index(min_confidence.value)
                 allowed_values = confidence_values[min_index:]
-                
+
                 filter_conditions["must"].append({
                     "key": "metadata.confidence",
                     "match": {"any": allowed_values}
                 })
-            
+
             # Additional custom filters
             if filter:
                 if "must" in filter:
                     filter_conditions["must"].extend(filter["must"])
                 else:
                     filter_conditions["must"].append(filter)
-            
+
             # Perform semantic search
             results = await super().search(query, limit, filter_conditions)
-            
+
             logger.info(f"Found {len(results)} semantic knowledge items matching query")
             logger.debug(f"Knowledge search results preview: {[r.id for r in results[:3]]}")
-            
+
             return results
-            
+
         except Exception as e:
             logger.error(f"Error searching semantic knowledge: {e}")
             return []
-    
+
     @log_function()
     async def search_sender_knowledge(
         self,
         sender_email: str,
         limit: int = 5
-    ) -> List[MemoryItem]:
+    ) -> list[MemoryItem]:
         """
         Search knowledge about a specific sender.
         
@@ -449,25 +446,25 @@ class SemanticMemory(BaseMemory):
             List of knowledge items about the specified sender
         """
         logger.info(f"Searching sender knowledge for: {sender_email}")
-        
+
         sender_filter = {
             "key": "metadata.sender_email",
             "match": {"value": sender_email}
         }
-        
+
         return await self.search(
             query=sender_email,
             limit=limit,
             filter=sender_filter,
             knowledge_type=KnowledgeType.SENDER
         )
-    
+
     @log_function()
     async def search_by_domain(
         self,
         domain: str,
         limit: int = 10
-    ) -> List[MemoryItem]:
+    ) -> list[MemoryItem]:
         """
         Search knowledge by domain or organization.
         
@@ -482,7 +479,7 @@ class SemanticMemory(BaseMemory):
             List of knowledge items related to the domain
         """
         logger.info(f"Searching domain knowledge for: {domain}")
-        
+
         domain_filter = {
             "must": [
                 {
@@ -494,15 +491,15 @@ class SemanticMemory(BaseMemory):
                 }
             ]
         }
-        
+
         return await self.search(
             query=domain,
             limit=limit,
             filter=domain_filter
         )
-    
+
     @log_function()
-    async def get_knowledge_statistics(self) -> Dict[str, Any]:
+    async def get_knowledge_statistics(self) -> dict[str, Any]:
         """
         Get complete statistics about semantic knowledge.
         
@@ -513,11 +510,11 @@ class SemanticMemory(BaseMemory):
             Dictionary containing knowledge statistics and insights
         """
         logger.info("Generating semantic knowledge statistics")
-        
+
         try:
             # Get all knowledge for analysis
             all_knowledge = await super().search(query="*", limit=1000)
-            
+
             if not all_knowledge:
                 return {
                     "total_count": 0,
@@ -525,7 +522,7 @@ class SemanticMemory(BaseMemory):
                     "confidence_distribution": {},
                     "average_content_length": 0
                 }
-            
+
             # Initialize statistics
             stats = {
                 "total_count": len(all_knowledge),
@@ -537,52 +534,52 @@ class SemanticMemory(BaseMemory):
                 "top_domains": {},
                 "top_organizations": {}
             }
-            
+
             # Analyze knowledge items
             total_length = 0
-            
+
             for knowledge in all_knowledge:
                 metadata = knowledge.metadata
-                
+
                 # Content length analysis
                 total_length += len(knowledge.content)
-                
+
                 # Knowledge type distribution
                 knowledge_type = metadata.get("knowledge_type", "unknown")
                 stats["knowledge_types"][knowledge_type] = stats["knowledge_types"].get(knowledge_type, 0) + 1
-                
+
                 # Confidence distribution
                 confidence = metadata.get("confidence", "unknown")
                 stats["confidence_distribution"][confidence] = stats["confidence_distribution"].get(confidence, 0) + 1
-                
+
                 # Sender and domain analysis
                 if knowledge_type == "sender":
                     stats["sender_knowledge_count"] += 1
-                    
+
                     sender_domain = metadata.get("sender_domain")
                     if sender_domain:
                         stats["top_domains"][sender_domain] = stats["top_domains"].get(sender_domain, 0) + 1
-                    
+
                     organization = metadata.get("organization")
                     if organization:
                         stats["top_organizations"][organization] = stats["top_organizations"].get(organization, 0) + 1
-                
+
                 elif knowledge_type == "domain":
                     stats["domain_knowledge_count"] += 1
-            
+
             # Calculate derived statistics
             stats["average_content_length"] = total_length / len(all_knowledge) if all_knowledge else 0
-            
+
             # Find most common items
             if stats["knowledge_types"]:
                 stats["most_common_type"] = max(stats["knowledge_types"], key=stats["knowledge_types"].get)
-            
+
             if stats["top_domains"]:
                 stats["most_documented_domain"] = max(stats["top_domains"], key=stats["top_domains"].get)
-            
+
             logger.info(f"Generated statistics for {stats['total_count']} knowledge items")
             return stats
-            
+
         except Exception as e:
             logger.error(f"Error generating knowledge statistics: {e}")
             return {
@@ -601,13 +598,13 @@ async def demo_semantic_memory() -> None:
     and knowledge search for asset management environments.
     """
     logger.info("Starting SemanticMemory demonstration")
-    
+
     semantic_memory = SemanticMemory()
-    
+
     try:
         # Add sample sender knowledge
         logger.info("Adding sample sender knowledge...")
-        
+
         sender_id = await semantic_memory.add_sender_knowledge(
             content="Responds within 2 hours during business hours, prefers detailed technical analysis in investment proposals",
             sender_email="investment@blackstone.com",
@@ -617,10 +614,10 @@ async def demo_semantic_memory() -> None:
             decision_authority="investment_committee",
             preferred_communication="technical_detailed"
         )
-        
+
         # Add email type knowledge
         logger.info("Adding email type knowledge...")
-        
+
         email_type_id = await semantic_memory.add_email_type_knowledge(
             content="Investment inquiries typically include fund size, target returns, and investment timeline",
             email_type="investment_inquiry",
@@ -628,10 +625,10 @@ async def demo_semantic_memory() -> None:
             typical_patterns=["fund size", "target return", "investment timeline", "due diligence"],
             confidence=KnowledgeConfidence.HIGH
         )
-        
+
         # Add domain knowledge
         logger.info("Adding domain expertise...")
-        
+
         domain_id = await semantic_memory.add_domain_knowledge(
             content="Real estate private equity typically requires 6-12 month due diligence periods for large transactions ($100M+)",
             domain="real_estate",
@@ -639,63 +636,63 @@ async def demo_semantic_memory() -> None:
             industry_context="commercial_real_estate",
             confidence=KnowledgeConfidence.HIGH
         )
-        
+
         logger.info("✓ Added sample semantic knowledge")
-        
+
         # Test knowledge search
         logger.info("Testing knowledge search capabilities...")
-        
+
         search_results = await semantic_memory.search(
             query="investment response time",
             knowledge_type=KnowledgeType.SENDER,
             limit=5
         )
-        
+
         logger.info(f"Found {len(search_results)} results for investment response time")
         for result in search_results:
             knowledge_type = result.metadata.get("knowledge_type", "unknown")
             confidence = result.metadata.get("confidence", "unknown")
             logger.info(f"  - {knowledge_type} ({confidence}): {result.content[:80]}...")
-        
+
         # Test sender-specific search
         logger.info("Testing sender-specific knowledge search...")
-        
+
         sender_results = await semantic_memory.search_sender_knowledge(
             sender_email="investment@blackstone.com"
         )
-        
+
         logger.info(f"Found {len(sender_results)} knowledge items about investment@blackstone.com")
-        
+
         # Test domain search
         logger.info("Testing domain knowledge search...")
-        
+
         domain_results = await semantic_memory.search_by_domain(
             domain="blackstone.com",
             limit=10
         )
-        
+
         logger.info(f"Found {len(domain_results)} knowledge items related to blackstone.com")
-        
+
         # Generate knowledge statistics
         logger.info("Generating knowledge statistics...")
-        
+
         stats = await semantic_memory.get_knowledge_statistics()
-        logger.info(f"Knowledge statistics:")
+        logger.info("Knowledge statistics:")
         logger.info(f"  - Total knowledge items: {stats['total_count']}")
         logger.info(f"  - Knowledge types: {stats['knowledge_types']}")
         logger.info(f"  - Confidence distribution: {stats['confidence_distribution']}")
         logger.info(f"  - Sender knowledge: {stats['sender_knowledge_count']}")
         logger.info(f"  - Domain knowledge: {stats['domain_knowledge_count']}")
-        
+
         if stats.get('most_common_type'):
             logger.info(f"  - Most common type: {stats['most_common_type']}")
-        
+
         logger.info("SemanticMemory demonstration completed successfully")
-        
+
     except Exception as e:
         logger.error(f"SemanticMemory demonstration failed: {e}")
         raise
 
 if __name__ == "__main__":
     import asyncio
-    asyncio.run(demo_semantic_memory()) 
+    asyncio.run(demo_semantic_memory())
