@@ -2,17 +2,18 @@
 """
 Startup script for the Email Agent Asset Management Web UI
 
-Run this script to start the web interface for managing assets, 
+Run this script to start the web interface for managing assets,
 sender mappings, and document classification settings.
 """
 
+# # Standard library imports
 import sys
-import os
 from pathlib import Path
 
 # Add src to path
 project_root = Path(__file__).parent
-sys.path.insert(0, str(project_root / 'src'))
+sys.path.insert(0, str(project_root / "src"))
+
 
 def main():
     """Main function to run the web UI"""
@@ -21,18 +22,20 @@ def main():
     print("Starting Flask development server...")
     print()
     print("📊 Dashboard:        http://localhost:5001")
-    print("🏢 Assets:           http://localhost:5001/assets") 
+    print("🏢 Assets:           http://localhost:5001/assets")
     print("📧 Sender Mappings:  http://localhost:5001/senders")
     print("🔧 API Health:       http://localhost:5001/api/health")
     print()
     print("Press Ctrl+C to stop the server")
     print("=" * 50)
-    
+
     # Import and run the Flask app
     try:
+        # # Local application imports
         from web_ui.app import create_app
+
         app = create_app()
-        app.run(debug=True, host='0.0.0.0', port=5001)
+        app.run(debug=True, host="0.0.0.0", port=5001)
     except ImportError as e:
         print(f"❌ Failed to import Flask app: {e}")
         print("Make sure Flask is installed: pip install Flask>=2.3.0")
@@ -41,5 +44,6 @@ def main():
         print(f"❌ Failed to start web UI: {e}")
         sys.exit(1)
 
-if __name__ == '__main__':
-    main() 
+
+if __name__ == "__main__":
+    main()
