@@ -14,4 +14,4 @@ Version: 1.0.0
 """
 
 __version__ = "1.0.0"
-__author__ = "Email Agent Development Team" 
+__author__ = "Email Agent Development Team"
